% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{bampath}
\alias{bampath}
\title{Access the path to a bam alignment file.}
\usage{
bampath(x, ...)
}
\arguments{
\item{x}{An object containing paths to bam files.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the path to a bam alignment file.
}
\examples{
###
}
