% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{check_alignment_file}
\alias{check_alignment_file}
\title{Manually check the alignment of consensus sequences using an editor.}
\usage{
check_alignment_file(x, which = "mapFinal", where = 0,
  editor = "xdg-open")
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{which}{Which .}
}
\value{
Side effect
}
\description{
Manually check the alignment of consensus sequences using an editor.
}
\details{
\code{report_checked_consensus} will create a \code{checked} directory
containing the consensus sequences \code{\{A,B\}.\{readtype\}.\{mapper\}.checked.fa}
and a html alignment file \code{aln.\{readtype\}.\{mapper\}.checked.html}.
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extractLongReads}}, \code{\link{mapFinal}},
  \code{\link{mapInit}}, \code{\link{mapIter}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}}, \code{\link{splitReadsByHaplotype}}
}
