% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequtils.R
\name{checkHomoPolymerCount}
\alias{checkHomoPolymerCount}
\title{Get a distribution of homopolymer count in alleles;}
\usage{
checkHomoPolymerCount(x, count = 8, map = "mapFinal")
}
\arguments{
\item{x}{a DR2S object}

\item{count}{the minimal length of a homopolymer to be found (10)}

\item{map}{Which result to use. Either "mapFinal" or "refine"}
}
\value{
plot a pdf with length histogram and write mode to log
}
\description{
Get a distribution of homopolymer count in alleles;
}
\examples{
###
}
