% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conseq.R
\name{conseq}
\alias{conseq}
\title{Construct a consensus sequence}
\usage{
conseq(x, name = "conseq", type = c("prob", "ambig", "simple"),
  threshold = NULL, suppressAllGaps = FALSE, suppressInsGaps = TRUE,
  gapThreshold = NULL, columnOccupancy = 0.4, ...)
}
\arguments{
\item{x}{\code{pileup} object.}

\item{name}{Name for consensus sequence.}

\item{type}{One of "prob", "ambig" or "simple".}

\item{threshold}{If \code{type = "ambig"}, threshold to call an ambiguous
consensus call.}

\item{suppressAllGaps}{If \code{type = "prob"} or \code{type = "ambig"},
suppress all gaps from consensus calling irrespective of the frequency of the gap.}

\item{suppressInsGaps}{If \code{type = "prob"}, suppress gaps at insertion position
from consensus calling based on \code{columnOccupancy}.}

\item{columnOccupancy}{Minimum occupancy (1 - fraction of gap) below which
bases at insertion position are excluded from from consensus calling.}

\item{...}{Additional arguments.}
}
\value{
A \code{\linkS4class{BStringSet}} object with a metadata list
containing the slots:
\describe{
  \item{zscore}{}
  \item{freq}{}
  \item{ambigs}{}
  \item{insertions}{}
  \item{deletions}{}
  \item{consmat}{}
}
}
\description{
Construct a consensus sequence
}
\examples{
###
}
