% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{consname}
\alias{consname}
\title{Access the consensus name.}
\usage{
consname(x, ...)
}
\arguments{
\item{x}{An object containing consensus information.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the consensus name.
}
\examples{
###
}
