% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{conspath}
\alias{conspath}
\title{Access the path to a consensus sequence file.}
\usage{
conspath(x, ...)
}
\arguments{
\item{x}{An object containing paths to consensus files.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the path to a consensus sequence file.
}
\examples{
###
}
