% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{createDR2SConf}
\alias{createDR2SConf}
\title{Constructor for configuration to create a \code{\link[=DR2S_]{DR2S}} object.}
\usage{
createDR2SConf(sample, locus, longreads = list(type = "pacbio", 
dir = "pacbio"), shortreads = list(type = "illumina", dir = "illumina"), 
datadir = ".", outdir = "./output", reference = NULL,
threshold = 0.20, iterations = 1, microsatellite = FALSE, distAlleles = 2, 
filterScores = TRUE, partSR = TRUE, forceMapping = FALSE, fullname = TRUE, 
details = NULL, ...)
}
\arguments{
\item{sample}{A unique sample identifier used to locate the long and short
read FASTQ files.}

\item{locus}{The HLA or KIR locus.}

\item{longreads}{The type and location of the long reads as a named list
with the fields \code{type} ("pacbio" or "nanopore") and \code{dir}.
Additional optional fields: \code{name} and \code{opts}.}

\item{shortreads}{The type and location of the short reads as a named list
with the fields \code{type} ("illumina") and \code{dir} (optional).}

\item{datadir}{The data directory (See Note).}

\item{outdir}{The output directory (See Note).}

\item{reference}{The reference allele(s).}

\item{threshold}{Threshold frequency for polymorphisms.}

\item{iterations}{Number of iterations of the mapIter step.}

\item{microsatellite}{FALSE Perform a second mapping of shortreads to the 
inferred reference in mapInit. Set to TRUE if you know you have repeats like 
in microsatellites. Usually extends the reference to a maximum length and 
enables a better mapping.}

\item{distAlleles}{Number of different alleles in the sample. Should be 2
for heterozygous samples, 1 for homozygous samples and > 2 for some KIR loci.}

\item{filterScores}{use only reads passing a strict filtering step. TODO}

\item{partSR}{Use shortreads in the mapInit step for getting polymorphic 
positions and a first reference.}

\item{forceMapping}{FALSE set to TRUE if you want to force processing of bad
shortreads, i.e. when the distribution of coverage is bad. Aborts the program
if maximum coverage > 75 \% quantile * 5.}

\item{fullname}{Truncate allele names.}

\item{details}{Metadata of a sample. Will be written to the fasta header of 
the final sequences and stored in the config dump.}

\item{...}{Additional arguments.}
}
\value{
A \code{\link[DR2S]{DR2S}} config object
}
\description{
Constructor for configuration to create a \code{\link[=DR2S_]{DR2S}} object.
}
\section{datadir}{


A \code{datadir} must contain arbitrarily named subdirectories
(default: "pacbio" and "illumina") with long read FASTQs in the format
\code{SAMPLE_LOCUS_*.fastq(.gz)} and short read FASTQs in the format
\code{SAMPLE_LOCUS_*R[12]*.fastq(.gz)}, respectively.
}

\section{Outdir}{


All output will be placed in directory hierarchy
\code{OUTDIR/SAMPLE}
}

\section{Reference}{


References can be specified as allele codes or a path to a fasta file
containing the reference sequence.
}

\examples{
\dontrun{
x <- InitDR2S(createDR2SConf(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  )) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  splitLongReads() \%>\%
  extractFastq() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{mapFinal}},
  \code{\link{mapInit}}, \code{\link{mapIter}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
