% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR2S-utils.R
\name{createIgvConfigs}
\alias{createIgvConfigs}
\title{Create IGV xml config files for directly open the longread and shortread
mapping in an IGV session. It creates also ".sh" and ".bat" scripts for 
easy opening the IGV instance.}
\usage{
createIgvConfigs(x, position, map = "mapFinal", open = TRUE)
}
\arguments{
\item{x}{A \code{\link{DR2S_}} object.}

\item{position}{The position where IGV focuses on startup.}

\item{map}{Which mapping should be opened. One of "mapInit", "mapIter", 
"mapFinal" or "refine".}

\item{open}{whether to open an IGV instance now.}
}
\description{
Create IGV xml config files for directly open the longread and shortread
mapping in an IGV session. It creates also ".sh" and ".bat" scripts for 
easy opening the IGV instance.
}
