% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consmat.R
\name{createPWM}
\alias{createPWM}
\title{Create position weight matrix from multiple sequence alignment}
\usage{
createPWM(msa)
}
\arguments{
\item{msa}{A \code{DNAStringSet} object of aligned sequences.}
}
\value{
A \code{PWM} matrix.
}
\description{
Create position weight matrix from multiple sequence alignment
}
\details{
\code{PWM}: a \code{matrix} with positions row names and nucleotides as
column manes. Values are nucleotide weights at a position
A ConsensusMatrix is calculated from the MSA using 
\code{Biostrings::consensusMatrix} and values are converted to probabilities.
Pseudocounts are added and values are divided by DNA probabilities and log2 
score is reported
}
\examples{
###
}
