% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeEnv-methods.R
\name{disambiguate_variant}
\alias{disambiguate_variant}
\title{Disambiguate a variant}
\usage{
disambiguate_variant(x, threshold, min_overrep_lr = 1.5, max_overrep_sr = 2)
}
\arguments{
\item{x}{A \code{variant_list}.}

\item{threshold}{Minimum frequency for a variant.}

\item{min_overrep_lr}{Issue a warning if the majority variant in the
long reads is only overrepresented by less than \code{min_overrep_lr}.}

\item{max_overrep_sr}{Issue a warning if the majority variant in the
short reads is overrepresented by more than \code{max_overrep_sr}.}
}
\value{
A \code{variant_list}.
}
\description{
Disambiguate a variant
}
\examples{
###
}
\keyword{internal}

