% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequtils.R
\name{.generateReferenceSequence}
\alias{.generateReferenceSequence}
\title{Generate reference sequences in a FASTA file}
\usage{
.generateReferenceSequence(
  locus,
  allele = NULL,
  outdir,
  utrLength = NULL,
  dirtag = NULL
)
}
\arguments{
\item{locus}{Locus name (HLA or KIR).}

\item{allele}{Allele name or NULL.}

\item{outdir}{Output directory.}

\item{dirtag}{TODO}
}
\value{
Path to refseq fasta file.
}
\description{
Generate reference sequences in a FASTA file
}
\examples{
###
}
\keyword{internal}
