% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR2S-class.R
\name{DR2S_}
\alias{DR2S_}
\title{R6 Class \code{"DR2S"}}
\description{
An object representing a DR2S analysis.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mapInit}}{<MapList>; the initial mapping of long reads to
\code{reference}.}

\item{\code{lrpartition}}{<PartList>; the partitioning of full-length
mapped long reads into different haplotypes.}

\item{\code{mapIter}}{<MapList>; the mapping of A and B reads to
consensus sequences produced in the previous step.}

\item{\code{srpartition}}{<PartList>; the partitioning of mapped short reads
into different haplotypes.}

\item{\code{mapFinal}}{<MapList>; the mapping of A, B, and short reads
to consensus sequences produced in the previous step.}

\item{\code{remap}}{<MapList>; the mapping of A, B, and short reads
to consensus sequences produced in the previous step.}

\item{\code{consensus}}{<ConsList>; final consensus sequences for A and
B.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DR2S_$new()}}
\item \href{#method-cache}{\code{DR2S_$cache()}}
\item \href{#method-clear}{\code{DR2S_$clear()}}
\item \href{#method-cleanup}{\code{DR2S_$cleanup()}}
\item \href{#method-print}{\code{DR2S_$print()}}
\item \href{#method-run_}{\code{DR2S_$run_()}}
\item \href{#method-getConfig}{\code{DR2S_$getConfig()}}
\item \href{#method-setConfig}{\code{DR2S_$setConfig()}}
\item \href{#method-getDatadir}{\code{DR2S_$getDatadir()}}
\item \href{#method-getOutdir}{\code{DR2S_$getOutdir()}}
\item \href{#method-getMicrosatellite}{\code{DR2S_$getMicrosatellite()}}
\item \href{#method-getForceMapping}{\code{DR2S_$getForceMapping()}}
\item \href{#method-getThreshold}{\code{DR2S_$getThreshold()}}
\item \href{#method-getDistAlleles}{\code{DR2S_$getDistAlleles()}}
\item \href{#method-getIterations}{\code{DR2S_$getIterations()}}
\item \href{#method-getLrdMapper}{\code{DR2S_$getLrdMapper()}}
\item \href{#method-getLrdType}{\code{DR2S_$getLrdType()}}
\item \href{#method-getLrdDir}{\code{DR2S_$getLrdDir()}}
\item \href{#method-getLongreads}{\code{DR2S_$getLongreads()}}
\item \href{#method-getSrdMapper}{\code{DR2S_$getSrdMapper()}}
\item \href{#method-getSrdType}{\code{DR2S_$getSrdType()}}
\item \href{#method-getSrdDir}{\code{DR2S_$getSrdDir()}}
\item \href{#method-getShortreads}{\code{DR2S_$getShortreads()}}
\item \href{#method-getPipeline}{\code{DR2S_$getPipeline()}}
\item \href{#method-getOpts}{\code{DR2S_$getOpts()}}
\item \href{#method-getSampleId}{\code{DR2S_$getSampleId()}}
\item \href{#method-getLocus}{\code{DR2S_$getLocus()}}
\item \href{#method-getReference}{\code{DR2S_$getReference()}}
\item \href{#method-getPlatform}{\code{DR2S_$getPlatform()}}
\item \href{#method-getFormat}{\code{DR2S_$getFormat()}}
\item \href{#method-setFormat}{\code{DR2S_$setFormat()}}
\item \href{#method-getDetails}{\code{DR2S_$getDetails()}}
\item \href{#method-setDetails}{\code{DR2S_$setDetails()}}
\item \href{#method-getSampleDetails}{\code{DR2S_$getSampleDetails()}}
\item \href{#method-getStats}{\code{DR2S_$getStats()}}
\item \href{#method-setStats}{\code{DR2S_$setStats()}}
\item \href{#method-getLimits}{\code{DR2S_$getLimits()}}
\item \href{#method-setLimits}{\code{DR2S_$setLimits()}}
\item \href{#method-getHapTypes}{\code{DR2S_$getHapTypes()}}
\item \href{#method-setHapTypes}{\code{DR2S_$setHapTypes()}}
\item \href{#method-getRefPath}{\code{DR2S_$getRefPath()}}
\item \href{#method-setRefPath}{\code{DR2S_$setRefPath()}}
\item \href{#method-getHomozygous}{\code{DR2S_$getHomozygous()}}
\item \href{#method-setHomozygous}{\code{DR2S_$setHomozygous()}}
\item \href{#method-getRefSeq}{\code{DR2S_$getRefSeq()}}
\item \href{#method-getPileup}{\code{DR2S_$getPileup()}}
\item \href{#method-getSnpMatrix}{\code{DR2S_$getSnpMatrix()}}
\item \href{#method-getPartition}{\code{DR2S_$getPartition()}}
\item \href{#method-getHapList}{\code{DR2S_$getHapList()}}
\item \href{#method-getLatestRefPath}{\code{DR2S_$getLatestRefPath()}}
\item \href{#method-getLatestRef}{\code{DR2S_$getLatestRef()}}
\item \href{#method-getMapTag}{\code{DR2S_$getMapTag()}}
\item \href{#method-getGroupPileup}{\code{DR2S_$getGroupPileup()}}
\item \href{#method-getLrdMapFun}{\code{DR2S_$getLrdMapFun()}}
\item \href{#method-getSrdMapFun}{\code{DR2S_$getSrdMapFun()}}
\item \href{#method-absPath}{\code{DR2S_$absPath()}}
\item \href{#method-relPath}{\code{DR2S_$relPath()}}
\item \href{#method-hasMapInit}{\code{DR2S_$hasMapInit()}}
\item \href{#method-hasLrdPartition}{\code{DR2S_$hasLrdPartition()}}
\item \href{#method-hasHapList}{\code{DR2S_$hasHapList()}}
\item \href{#method-hasMapIter}{\code{DR2S_$hasMapIter()}}
\item \href{#method-hasMapFinal}{\code{DR2S_$hasMapFinal()}}
\item \href{#method-checkGetShortreads}{\code{DR2S_$checkGetShortreads()}}
\item \href{#method-hasShortreads}{\code{DR2S_$hasShortreads()}}
\item \href{#method-checkGetLongreads}{\code{DR2S_$checkGetLongreads()}}
\item \href{#method-hasLongreads}{\code{DR2S_$hasLongreads()}}
\item \href{#method-plotCoverage}{\code{DR2S_$plotCoverage()}}
\item \href{#method-plotGroupCoverage}{\code{DR2S_$plotGroupCoverage()}}
\item \href{#method-plotBasecallFrequency}{\code{DR2S_$plotBasecallFrequency()}}
\item \href{#method-plotGroupBasecallFrequency}{\code{DR2S_$plotGroupBasecallFrequency()}}
\item \href{#method-plotPartitionHistogram}{\code{DR2S_$plotPartitionHistogram()}}
\item \href{#method-plotPartitionTree}{\code{DR2S_$plotPartitionTree()}}
\item \href{#method-plotPartitionRadar}{\code{DR2S_$plotPartitionRadar()}}
\item \href{#method-polymorphicPositions}{\code{DR2S_$polymorphicPositions()}}
\item \href{#method-plotMapInitSummary}{\code{DR2S_$plotMapInitSummary()}}
\item \href{#method-plotPartitionSummary}{\code{DR2S_$plotPartitionSummary()}}
\item \href{#method-plotMapIterSummary}{\code{DR2S_$plotMapIterSummary()}}
\item \href{#method-plotMapFinalSummary}{\code{DR2S_$plotMapFinalSummary()}}
\item \href{#method-plotRemapSummary}{\code{DR2S_$plotRemapSummary()}}
\item \href{#method-plotSeqLogo}{\code{DR2S_$plotSeqLogo()}}
\item \href{#method-runMapInit}{\code{DR2S_$runMapInit()}}
\item \href{#method-runPartitionLongreads}{\code{DR2S_$runPartitionLongreads()}}
\item \href{#method-runSplitLongreadsByHaplotype}{\code{DR2S_$runSplitLongreadsByHaplotype()}}
\item \href{#method-runExtractPartitionedLongreads}{\code{DR2S_$runExtractPartitionedLongreads()}}
\item \href{#method-runMapIter}{\code{DR2S_$runMapIter()}}
\item \href{#method-runPartitionShortreads}{\code{DR2S_$runPartitionShortreads()}}
\item \href{#method-runMapFinal}{\code{DR2S_$runMapFinal()}}
\item \href{#method-runPipeline}{\code{DR2S_$runPipeline()}}
\item \href{#method-clone}{\code{DR2S_$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new DR2S object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$new(conf, createOutdir = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conf}}{A DR2S configuration provided either by [createDR2SConf] or
read from file by [readDR2SConf].}

\item{\code{createOutdir}}{Set up the output directory.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DR2S object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cache"></a>}}
\subsection{Method \code{cache()}}{
Store a DR2S object as an rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$cache(outname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outname}}{(optional) filename of the stored object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns invisibly the DR2S object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\subsection{Method \code{clear()}}{
Clear all the files pertaining to a DR2S mapping process
and set up the whole process from scratch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$clear()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns invisibly the DR2S object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanup"></a>}}
\subsection{Method \code{cleanup()}}{
remove all bam and bai files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$cleanup()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns invisibly the DR2S object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print a DR2S object prettily
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns invisibly the DR2S object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_"></a>}}
\subsection{Method \code{run_()}}{
Execute a given step in a DR2S pipeline.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$run_(step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{A step in the DR2S pipeline}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Called for side effect.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConfig"></a>}}
\subsection{Method \code{getConfig()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getConfig(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setConfig"></a>}}
\subsection{Method \code{setConfig()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$setConfig(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDatadir"></a>}}
\subsection{Method \code{getDatadir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getDatadir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOutdir"></a>}}
\subsection{Method \code{getOutdir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getOutdir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMicrosatellite"></a>}}
\subsection{Method \code{getMicrosatellite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getMicrosatellite()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getForceMapping"></a>}}
\subsection{Method \code{getForceMapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getForceMapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getThreshold"></a>}}
\subsection{Method \code{getThreshold()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getThreshold()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDistAlleles"></a>}}
\subsection{Method \code{getDistAlleles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getDistAlleles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIterations"></a>}}
\subsection{Method \code{getIterations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getIterations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLrdMapper"></a>}}
\subsection{Method \code{getLrdMapper()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLrdMapper()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLrdType"></a>}}
\subsection{Method \code{getLrdType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLrdType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLrdDir"></a>}}
\subsection{Method \code{getLrdDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLrdDir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLongreads"></a>}}
\subsection{Method \code{getLongreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLongreads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSrdMapper"></a>}}
\subsection{Method \code{getSrdMapper()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getSrdMapper()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSrdType"></a>}}
\subsection{Method \code{getSrdType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getSrdType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSrdDir"></a>}}
\subsection{Method \code{getSrdDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getSrdDir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getShortreads"></a>}}
\subsection{Method \code{getShortreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getShortreads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPipeline"></a>}}
\subsection{Method \code{getPipeline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getPipeline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOpts"></a>}}
\subsection{Method \code{getOpts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getOpts(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSampleId"></a>}}
\subsection{Method \code{getSampleId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getSampleId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLocus"></a>}}
\subsection{Method \code{getLocus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLocus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReference"></a>}}
\subsection{Method \code{getReference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getReference()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPlatform"></a>}}
\subsection{Method \code{getPlatform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getPlatform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFormat"></a>}}
\subsection{Method \code{getFormat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getFormat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setFormat"></a>}}
\subsection{Method \code{setFormat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$setFormat(format)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDetails"></a>}}
\subsection{Method \code{getDetails()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getDetails(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDetails"></a>}}
\subsection{Method \code{setDetails()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$setDetails(details)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSampleDetails"></a>}}
\subsection{Method \code{getSampleDetails()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getSampleDetails()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStats"></a>}}
\subsection{Method \code{getStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getStats(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setStats"></a>}}
\subsection{Method \code{setStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$setStats(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLimits"></a>}}
\subsection{Method \code{getLimits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLimits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLimits"></a>}}
\subsection{Method \code{setLimits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$setLimits(lmts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHapTypes"></a>}}
\subsection{Method \code{getHapTypes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getHapTypes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setHapTypes"></a>}}
\subsection{Method \code{setHapTypes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$setHapTypes(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRefPath"></a>}}
\subsection{Method \code{getRefPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getRefPath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setRefPath"></a>}}
\subsection{Method \code{setRefPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$setRefPath(refPath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHomozygous"></a>}}
\subsection{Method \code{getHomozygous()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getHomozygous()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setHomozygous"></a>}}
\subsection{Method \code{setHomozygous()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$setHomozygous(homozygous)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRefSeq"></a>}}
\subsection{Method \code{getRefSeq()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getRefSeq()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPileup"></a>}}
\subsection{Method \code{getPileup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getPileup(useSR = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSnpMatrix"></a>}}
\subsection{Method \code{getSnpMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getSnpMatrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPartition"></a>}}
\subsection{Method \code{getPartition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getPartition()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHapList"></a>}}
\subsection{Method \code{getHapList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getHapList(group)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLatestRefPath"></a>}}
\subsection{Method \code{getLatestRefPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLatestRefPath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLatestRef"></a>}}
\subsection{Method \code{getLatestRef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLatestRef()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMapTag"></a>}}
\subsection{Method \code{getMapTag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getMapTag(iter = "init", group = NULL, ref = "LR")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGroupPileup"></a>}}
\subsection{Method \code{getGroupPileup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getGroupPileup(group = NULL, ref = NULL, iteration = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLrdMapFun"></a>}}
\subsection{Method \code{getLrdMapFun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getLrdMapFun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSrdMapFun"></a>}}
\subsection{Method \code{getSrdMapFun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$getSrdMapFun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-absPath"></a>}}
\subsection{Method \code{absPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$absPath(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-relPath"></a>}}
\subsection{Method \code{relPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$relPath(filepath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasMapInit"></a>}}
\subsection{Method \code{hasMapInit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$hasMapInit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasLrdPartition"></a>}}
\subsection{Method \code{hasLrdPartition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$hasLrdPartition()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasHapList"></a>}}
\subsection{Method \code{hasHapList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$hasHapList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasMapIter"></a>}}
\subsection{Method \code{hasMapIter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$hasMapIter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasMapFinal"></a>}}
\subsection{Method \code{hasMapFinal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$hasMapFinal()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkGetShortreads"></a>}}
\subsection{Method \code{checkGetShortreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$checkGetShortreads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasShortreads"></a>}}
\subsection{Method \code{hasShortreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$hasShortreads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkGetLongreads"></a>}}
\subsection{Method \code{checkGetLongreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$checkGetLongreads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasLongreads"></a>}}
\subsection{Method \code{hasLongreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$hasLongreads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCoverage"></a>}}
\subsection{Method \code{plotCoverage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotCoverage(
  threshold,
  range = NULL,
  thin = 0.1,
  width = 1,
  label = "",
  drop.indels = FALSE,
  readtype = "LR"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGroupCoverage"></a>}}
\subsection{Method \code{plotGroupCoverage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotGroupCoverage(
  group = NULL,
  ref = NULL,
  iteration = "init",
  threshold = NULL,
  range = NULL,
  thin = 0.1,
  width = 1,
  label = "",
  drop.indels = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotBasecallFrequency"></a>}}
\subsection{Method \code{plotBasecallFrequency()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotBasecallFrequency(threshold, label = "", drop.indels = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGroupBasecallFrequency"></a>}}
\subsection{Method \code{plotGroupBasecallFrequency()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotGroupBasecallFrequency(
  group = NULL,
  ref = NULL,
  iteration = "init",
  threshold = NULL,
  label = "",
  drop.indels = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPartitionHistogram"></a>}}
\subsection{Method \code{plotPartitionHistogram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotPartitionHistogram(label = "", limits = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPartitionTree"></a>}}
\subsection{Method \code{plotPartitionTree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotPartitionTree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPartitionRadar"></a>}}
\subsection{Method \code{plotPartitionRadar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotPartitionRadar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-polymorphicPositions"></a>}}
\subsection{Method \code{polymorphicPositions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$polymorphicPositions(threshold, useSR = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMapInitSummary"></a>}}
\subsection{Method \code{plotMapInitSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotMapInitSummary(thin = 0.2, width = 4, label = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPartitionSummary"></a>}}
\subsection{Method \code{plotPartitionSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotPartitionSummary(label = "", limits = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMapIterSummary"></a>}}
\subsection{Method \code{plotMapIterSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotMapIterSummary(
  thin = 0.2,
  width = 10,
  iteration = 0,
  drop.indels = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMapFinalSummary"></a>}}
\subsection{Method \code{plotMapFinalSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotMapFinalSummary(readtype, thin = 0.2, width = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotRemapSummary"></a>}}
\subsection{Method \code{plotRemapSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotRemapSummary(readtype, thin = 0.2, width = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSeqLogo"></a>}}
\subsection{Method \code{plotSeqLogo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$plotSeqLogo(ppos = NULL, pwm = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runMapInit"></a>}}
\subsection{Method \code{runMapInit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$runMapInit(opts = list(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runPartitionLongreads"></a>}}
\subsection{Method \code{runPartitionLongreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$runPartitionLongreads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runSplitLongreadsByHaplotype"></a>}}
\subsection{Method \code{runSplitLongreadsByHaplotype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$runSplitLongreadsByHaplotype()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runExtractPartitionedLongreads"></a>}}
\subsection{Method \code{runExtractPartitionedLongreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$runExtractPartitionedLongreads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runMapIter"></a>}}
\subsection{Method \code{runMapIter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$runMapIter(opts = list(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runPartitionShortreads"></a>}}
\subsection{Method \code{runPartitionShortreads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$runPartitionShortreads(opts = list(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runMapFinal"></a>}}
\subsection{Method \code{runMapFinal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$runMapFinal(opts = list(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runPipeline"></a>}}
\subsection{Method \code{runPipeline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$runPipeline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR2S_$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
