% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR2S-package.R
\docType{package}
\name{DR2Sdevel}
\alias{DR2Sdevel}
\alias{DR2Sdevel-package}
\title{DR2S - Dual redundant reference sequencing.}
\description{
Generate highly accurate reference sequences from heterozygous
HLA and KIR samples by combining data from long-read and short-read
sequencing technologies
}
