% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{DR2Smap}
\alias{DR2Smap}
\title{Constructor for \code{\link[=DR2S_]{DR2S}} mapper objects.}
\usage{
DR2Smap(sample, locus, longreads = list(type = "pacbio", dir = "pacbio"),
  shortreads = list(type = "illumina", dir = "illumina"), datadir = ".",
  outdir = "./output", reference = NULL, consensus = "multialign",
  threshold = 0.2, fullname = TRUE, ...)
}
\arguments{
\item{sample}{A unique sample identifier used to locate the long and short
read FASTQ files.}

\item{locus}{The HLA locus ("A", "B", "C", "DPB1", "DQB1", "DRB1")}

\item{longreads}{The type and location of the long reads as a named list
with the fields \code{type} ("pacbio" or "nanopore") and \code{dir}.
Additional optional fields: \code{name} and \code{opts}.}

\item{shortreads}{The type and location of the short reads as a named list
with the fields \code{type} ("illumina") and \code{dir}.}

\item{datadir}{The data directory (See Note).}

\item{outdir}{The output directory (See Note).}

\item{reference}{(Optional) The reference allele(s). One or both alleles
specified as a character vector (See Note).}

\item{consensus}{How will the inital allele consensus sequences be generated
from the long reads. One of \dQuote{\code{multialign}} or \dQuote{\code{mapping}}
(See Note).}

\item{threshold}{Threshold frequency for polymorphisms.}

\item{fullname}{Truncate allele names.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Initialise a \code{\link[=DR2S_]{DR2S}} mapper.
}
\section{datadir}{


A \code{datadir} must contain arbitrarily named subdirectories
(default: "pacbio" and "illumina") with long read FASTQs in the format
\code{SAMPLE_LOCUS_*.fastq(.gz)} and short read FASTQs in the format
\code{SAMPLE_LOCUS_*R[12]*.fastq(.gz)}, respectively.
}

\section{Outdir}{


All output will be placed in directory hierarchy
\code{OUTDIR/SAMPLE/READTYPE/LOCUS}.
}

\section{Reference}{


References can be specified as allele codes or a path to a fasta file
containing the reference sequence. Specifying both alleles as a character
vector is possible but only takes effect if \code{consensus} is set to
\dQuote{\code{mapping}}. If \code{reference = NULL} a global generic reference
for a given locus will be used.
}

\section{Consensus}{

Two methods to generate initial consensus sequences from long reads are
available: \dQuote{\bold{\code{multialign}}}: After partitioning long reads into
haplotypes up to 30 "best" reads from each haplotype group are selected for
multiple sequence alignment and a preliminary consensus sequences construction.
All reads from each haplotype group are subsequently mapped against their
respective preliminary consensus sequences and these mappings are used for
constructing final long read consensus sequences.
\dQuote{\bold{\code{mapping}}}: Reads from each haplotype group are mapped
against the orignal reference sequence and these mappings are used for
constructing final long read consensus sequences. Note, that the second
approach is prone to introducing alignment artifacts if there is a lot of
mismatch between the reference used and the actual allele sequences.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "multialign"
  ) \%>\%
  clear() \%>\%
  map0() \%>\%
  partition_haplotypes() \%>\%
  split_reads_by_haplotype() \%>\%
  extract_fastq() \%>\%
  map1() \%>\%
  map2() \%>\%
  map3() \%>\%
  cache() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S.mapper.functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{extract_fastq}},
  \code{\link{map0}}, \code{\link{map1}},
  \code{\link{map2}}, \code{\link{map3}},
  \code{\link{partition_haplotypes}}, \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}},
  \code{\link{split_reads_by_haplotype}}
}

