% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequtils.R
\name{extract_fastq.default}
\alias{extract_fastq.default}
\title{Extract (a subset of) reads from a bamfile}
\usage{
\method{extract_fastq}{default}(x, qnames = NULL, n = 100, replace = TRUE)
}
\arguments{
\item{x}{Path to bamfile.}

\item{qnames}{Sample a fixed subset of reads.}

\item{n}{Sample a fixed number of n reads.}

\item{replace}{Sample with or without replacemant?}
}
\value{
A \code{\linkS4class{ShortReadQ}} object
}
\description{
Extract (a subset of) reads from a bamfile
}
\examples{
###
}

