% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequtils.R
\name{filter_bam}
\alias{filter_bam}
\title{Filter a bamfile by qnames}
\usage{
filter_bam(x, qnames, dest)
}
\arguments{
\item{x}{Path to bamfile.}

\item{qnames}{Sample a fixed subset of reads.}

\item{dest}{Location where the filtered output file will be created.}
}
\value{
A character vector
}
\description{
Filter a bamfile by qnames
}
\examples{
###
}
