% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequtils.R
\name{generate_reference_sequence}
\alias{generate_reference_sequence}
\title{Generate reference sequences in a FASTA file}
\usage{
generate_reference_sequence(hla, allele, outdir, fullname = TRUE)
}
\arguments{
\item{allele}{Allele name.}

\item{outdir}{Output directory}

\item{fullname}{Truncate allele names}

\item{HLA}{A \code{HLAGene} object}
}
\value{
Path to refseq fasta file.
}
\description{
Generate reference sequences in a FASTA file
}
\examples{
###
}
\keyword{internal}
