% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionShortReads.R
\name{get_SR_partition_scores}
\alias{get_SR_partition_scores}
\title{Get a score for each shortread in a bam file on specific positions.
Scores based on a consensus matrix of longreads.}
\usage{
get_SR_partition_scores(refname, bamfile, mats, cores = "auto")
}
\arguments{
\item{refname}{Name of the reference.}

\item{bamfile}{BAM file path.}

\item{mats}{Consensus matrix from longread clustering at positions in ppos.}

\item{cores}{Number of cores or "auto"}

\item{ppos}{List of polymorphic positions used for initial clustering.}
}
\value{
A \code{srpatition} object. See \code{Details}.
}
\description{
Get a score for each shortread in a bam file on specific positions.
Scores based on a consensus matrix of longreads.
}
\details{
Returns a \code{srpartition} object:
A \code{list} with slots:
\describe{
  \item{bamfile}{<character>; Path to the bam file used to construct the pileup}
  \item{ppos}{<numeric>; Positions used for scoring}
  \item{srpartition}{A \code{data.frame} with colums:
    \describe{
      \item{read}{<character>; The read name}
      \item{pos}{<integer>; Genomic position of base}
      \item{Haplotype}{<character>; Haplotype}
      \item{prob}{<integer>; probability for a base at a position}
    }
  }
  \item{consmat}{Consensus matrix}
}
}
\examples{
###
}
