% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionShortReads.R
\name{getSRPartitionScores}
\alias{getSRPartitionScores}
\title{Get a score for each shortread in a bam file on specific positions.
Scores based on a consensus matrix of longreads.}
\usage{
getSRPartitionScores(bamfile, mats, ...)
}
\arguments{
\item{bamfile}{BAM file path.}

\item{mats}{Consensus matrix from longread clustering at positions in ppos.}
}
\value{
A \code{srpatition} object. See \code{Details}.
}
\description{
Get a score for each shortread in a bam file on specific positions.
Scores based on a consensus matrix of longreads.
}
\details{
Returns a \code{srpartition} object:
A \code{list} with slots:
\describe{
  \item{bamfile}{<character>; Path to the bam file used to construct the
  pileup}
  \item{ppos}{<numeric>; Positions used for scoring}
  \item{srpartition}{A \code{data.frame} with colums:
    \describe{
      \item{read}{<character>; The read name}
      \item{pos}{<integer>; Genomic position of base}
      \item{Haplotype}{<character>; Haplotype}
      \item{prob}{<integer>; probability for a base at a position}
    }
  }
  \item{consmat}{Consensus matrix}
}
}
\examples{
###
}
