% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{hammingDist}
\alias{hammingDist}
\title{Construct a hamming distance matrix from an aligned DNAStringSet}
\usage{
hammingDist(x)
}
\arguments{
\item{x}{DNAStringset with aligned sequences}
}
\value{
A \code{consmat} object.
}
\description{
Construct a hamming distance matrix from an aligned DNAStringSet
}
\details{
\code{consmat}: a \code{matrix} with positions row names and nucleotides as
column manes.
A \code{consmat} object includes the attributes:
\describe{
  \item{n}{<integer>; Number of reads per row.}
  \item{freq}{<logical>; Is it a frequency matrix or a count matrix.}
  \item{offset}{<integer>; Offset}
  \item{insertions}{<integer>; Insertions}
}
}
\examples{
###
}
