% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish-utils.R
\name{inspect_variant}
\alias{inspect_variant}
\title{Print diagnostic summary of a variant}
\usage{
inspect_variant(x, i, j = i)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} mapper object}

\item{i}{Variant index for haplotype A}

\item{j}{Variant index for haplotype B}
}
\description{
Print diagnostic summary of a variant
}

