% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{map2}
\alias{map2}
\title{Third mapping step (long reads).}
\usage{
map2(x, opts = list(), pct = 100, min_base_quality = 7, min_mapq = 0,
  max_depth = 10000, min_nucleotide_depth = 3, include_deletions = TRUE,
  include_insertions = TRUE, pruning_cutoff = 0.75, force = FALSE,
  fullname = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{pct}{Percentage of templates to subsample (before sorting the bam file).}

\item{min_base_quality}{Minimum \sQuote{QUAL} value for each nucleotide in an
alignment.}

\item{min_mapq}{Minimum \sQuote{MAPQ} value for an alignment to be included
in pileup.}

\item{max_depth}{Maximum number of overlapping alignments considered for
each position in the pileup.}

\item{min_nucleotide_depth}{Minimum count of each nucleotide at a given
position required for that nucleotide to appear in the result.}

\item{include_deletions}{If \code{TRUE}, include deletions in pileup.}

\item{include_insertions}{If \code{TRUE}, include insertions in pileup.}

\item{pruning_cutoff}{TODO}

\item{force}{If \code{TRUE}, overwrite existing bam file.}

\item{fullname}{If \code{TRUE}, use the full name in reference fasta.}

\item{plot}{Plot diagnostics.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Partitioned long reads are mapped against the consensus sequences inferred
at \code{map1}. Indels are \strong{included} in pileup. New consensus
sequences for both alleles are inferred.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "multialign"
  ) \%>\%
  clear() \%>\%
  map0() \%>\%
  partition_haplotypes() \%>\%
  split_reads_by_haplotype(limitA = 0.5, limitB = 0.25) \%>\%
  extract_fastq() \%>\%
  map1() \%>\%
  map2() \%>\%
  map3() \%>\%
  cache() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S.mapper.functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extract_fastq}}, \code{\link{map0}},
  \code{\link{map1}}, \code{\link{map3}},
  \code{\link{partition_haplotypes}}, \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}},
  \code{\link{split_reads_by_haplotype}}
}

