% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeEnv-class.R
\name{MergeEnv}
\alias{MergeEnv}
\title{Constructor for \code{\link[=MergeEnv_]{MergeEnv}} objects.}
\usage{
MergeEnv(x, threshold)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{threshold}{When do we call a variant a variant.}
}
\value{
A \code{\link[=MergeEnv_]{MergeEnv}} object.
}
\description{
Constructor for \code{\link[=MergeEnv_]{MergeEnv}} objects.
}
\examples{
###
}
