% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeEnv-class.R
\docType{class}
\name{MergeEnv_}
\alias{MergeEnv_}
\title{Class \code{"MergeEnv"}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MergeEnv(x)
}
\value{
Object of \code{\link{R6Class}} representing a MergeEnv.
}
\description{
Class \code{"MergeEnv"}
}
\section{Fields}{

\describe{
\item{\code{threshold}}{\code{[numeric]}; when do we call a variant a variant.}

\item{\code{a}}{\code{[HapEnv, environment]}; container for all things haplotype A.}

\item{\code{b}}{\code{[HapEnv, environment]}; container for all things haplotype B.}

\item{\code{x}}{\code{[\link[=DR2S_]{DR2S}]}; the original \code{DR2S} object.}
}}

\section{Public Methods}{

\describe{
\item{\code{x$init(hapEnv)}}{Intialise a \code{HapEnv}}
\item{\code{x$walk(hapEnv, until = FALSE, verbose = FALSE)}}{Perform a SNP
walk along the consensus matrices until there are no SNPs to be walked
\code{until = FALSE}, until a certain position (\code{until = 1234}), or
one step at a time (\code{until = NULL}).}
}
}

\keyword{data}
\keyword{internal}
