% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{meta}
\alias{meta}
\title{Access the metadata of an object.}
\usage{
meta(x, name = NULL, ...)
}
\arguments{
\item{x}{An object containing metadata}

\item{name}{(Optional); slotname.}

\item{...}{Additional arguments.}
}
\value{
A named list.
}
\description{
Access the metadata of an object.
}
\examples{
###
}
