% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects.}

\item{plotlist}{A list of ggplot objects.}

\item{cols}{Number of columns in layout.}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
ggplot objects can be passed in \code{...}, or to \code{plotlist} (as a list of ggplot objects)
}
\details{
If the layout is something like \code{matrix(c(1, 2, 3, 3), nrow = 2, byrow = TRUE)},
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\references{
\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
}
\author{
Winston Chang
}
\keyword{internal}
