% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-reads.R
\name{partitionReads}
\alias{partitionReads}
\title{Cluster a SNP matrix into haplotype groups}
\usage{
partitionReads(x, clMethod = "ward.D", minLen = 0.5, skipGapFreq = 2/3,
  deepSplit = 1, threshold = 0.2, distAlleles = 2, sortBy = "count")
}
\arguments{
\item{x}{A \code{[read x position]} SNP matrix.}

\item{clMethod}{clustering method passed to \code{\link[stats]{hclust}}.}

\item{minLen}{Minimal fraction of SNPs that need to be covered by a read
to be used}

\item{skipGapFreq}{Skip a badly behaved polymorphic position if the
frequency of gaps within a haplotype group exceeds \code{skipGapFreq}.}

\item{deepSplit}{sensitivity parameter passed to
\code{\link[dynamicTreeCut]{cutreeHybrid}}.}

\item{threshold}{Only gaps above threshold will be treated as gaps.
Removes noisy positions in longreads.}
}
\value{
A \code{HapPart} object.
}
\description{
Cluster a SNP matrix into haplotype groups
}
\examples{
###
}
