% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{partitionShortReads}
\alias{partitionShortReads}
\title{Assign short reads from mapInit to haplotypes}
\usage{
partitionShortReads(x, force = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{force}{force the creation of new fastq files if they already exist}

\item{plot}{Produce diagnostic plots.}

\item{...}{Further arguments passed to methods.}

\item{shuffle}{Randomly shuffle polymorphic positions.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Assign short reads from mapInit to haplotypes
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  ) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partition_haplotypes() \%>\%
  split_reads_by_haplotype() \%>\%
  extract_fastq() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extract_fastq}}, \code{\link{mapFinal}},
  \code{\link{mapInit}}, \code{\link{mapIter}},
  \code{\link{partition_haplotypes}}, \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}},
  \code{\link{split_reads_by_haplotype}}
}
