% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{path}
\alias{path}
\title{Access the path of an object}
\usage{
path(x, ...)
}
\arguments{
\item{x}{An object containing paths.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the path of an object
}
\examples{
###
}
