% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup.R
\name{Pileup}
\alias{Pileup}
\title{Calculate pile-up for a BAM file.}
\usage{
Pileup(bamfile, threshold = 0.2, maxDepth = 250, minBaseQuality = 0,
  minMapq = 0, minNucleotideDepth = 1, minMinorAlleleDepth = 0,
  distinguishStrands = FALSE, distinguishNucleotides = TRUE,
  ignoreQueryNs = TRUE, includeDeletions = TRUE,
  includeInsertions = FALSE, leftBins = NULL, queryBins = NULL)
}
\arguments{
\item{bamfile}{BAM file path.}

\item{threshold}{Threshold used for SNP calling.}
}
\value{
A \code{pileup} object. See \code{Details}.
}
\description{
Calculate pile-up for a BAM file.
}
\details{
Returns a \code{pileup} object:
A \code{list} with slots:
\describe{
  \item{bamfile}{<character>; Path to the bam file used to construct the pileup}
  \item{threshold}{<numeric>; Threshold used for calling SNPs}
  \item{param}{A \code{\link[Rsamtools]{PileupParam}} object}
  \item{pilup}{A \code{data.frame} with colums:
    \describe{
      \item{seqnames}{<character>; The RNAME field}
      \item{pos}{<integer>; Genomic position of base}
      \item{nucleotide}{<character>; Base}
      \item{count}{<integer>; Count of base}
    }
  }
  \item{consmat}{Consensus matrix}
}
}
\examples{
###
}
