% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conseq.R
\name{plot_conseq_probability}
\alias{plot_conseq_probability}
\title{Plot probability of consensus bases.}
\usage{
plot_conseq_probability(cseqs, threshold = "auto", text_size = 3,
  point_size = 1)
}
\arguments{
\item{text_size}{Text size.}

\item{point_size}{Point size.}

\item{seqA}{consensus sequence A.}

\item{seqB}{consensus sequence B.}

\item{labelA}{Label A.}

\item{labelB}{Label B.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot probability of consensus bases.
}
\examples{
##
}
