% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-reads.R
\name{plot_partition_haplotypes}
\alias{plot_partition_haplotypes}
\title{Tile plot of haplotyped reads.}
\usage{
plot_partition_haplotypes(x, thin = 1, label = "", sort = TRUE,
  name_reads = FALSE)
}
\arguments{
\item{x}{A \code{HapPart} object.}

\item{thin}{Subsample reads.}

\item{label}{Optional plot label.}

\item{sort}{Sort}

\item{name_reads}{Name reads}
}
\value{
A \code{ggplot} object.
}
\description{
Tile plot of haplotyped reads.
}
\examples{
###
}
