% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-reads.R
\name{plot_partition_histogram_multi}
\alias{plot_partition_histogram_multi}
\title{Plot distribution of haplotype partitioned reads for more than two haplotypes}
\usage{
plot_partition_histogram_multi(x, limits, label = "")
}
\arguments{
\item{x}{A \code{HapPart} object.}

\item{label}{Optional plot label.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot distribution of haplotype partitioned reads for more than two haplotypes
}
\examples{
###

}
