% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-reads.R
\name{plot_partition_tree}
\alias{plot_partition_tree}
\title{Plot tree of initial clustering}
\usage{
plot_partition_tree(x)
}
\arguments{
\item{x}{A \code{HapPart} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot tree of initial clustering
}
\examples{
###
}
