% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conseq.R
\name{plotConseqProbability}
\alias{plotConseqProbability}
\title{Plot probability of consensus bases.}
\usage{
plotConseqProbability(cseqs, threshold = "auto", textSize = 3,
  pointSize = 1)
}
\arguments{
\item{textSize}{Text size.}

\item{pointSize}{Point size.}

\item{seqA}{consensus sequence A.}

\item{seqB}{consensus sequence B.}

\item{labelA}{Label A.}

\item{labelB}{Label B.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot probability of consensus bases.
}
\examples{
##
}
