% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotDiagnosticAlignment}
\alias{plotDiagnosticAlignment}
\title{Plot an alignment of all intermediate sequences and reference}
\usage{
plotDiagnosticAlignment(x, onlyFinal = FALSE)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{onlyFinal}{restrict to final sequences}
}
\description{
Plot an alignment of all intermediate sequences and reference
}
\details{
Calls \code{\link[DECIPHER]{BrowseSeqs}} to open the alignment of 
references, final and intermediate (only if \code{onlyFinal} is FALSE) in a 
browser. The alignment is created with \code{\link[DECIPHER]{AlignSeqs}}.
}
\examples{
###
}
