% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionLongReads.R
\name{plotPartitionHaplotypes}
\alias{plotPartitionHaplotypes}
\title{Tile plot of haplotyped reads.}
\usage{
plotPartitionHaplotypes(x, thin = 1, label = "", sort = TRUE,
  nameReads = FALSE)
}
\arguments{
\item{x}{A \code{HapPart} object.}

\item{thin}{Subsample reads. [1]}

\item{label}{Optional plot label.}

\item{sort}{Sort [TRUE]}

\item{nameReads}{Name reads [FALSE]}
}
\value{
A \code{ggplot} object.
}
\description{
Tile plot of haplotyped reads.
}
\examples{
###
}
