% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-reads.R
\name{plotPartitionHistogram}
\alias{plotPartitionHistogram}
\title{Plot distribution of haplotype partitioned reads}
\usage{
plotPartitionHistogram(x, label = "", limits = NULL)
}
\arguments{
\item{x}{A \code{HapPart} object.}

\item{label}{Optional plot label.}

\item{limits}{Manually provided limits for plotting. Defaults to read limits
from the \code{HapPart} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot distribution of haplotype partitioned reads
}
\examples{
###
}
