% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionLongReads.R
\name{plotPartitionHistogramMulti}
\alias{plotPartitionHistogramMulti}
\title{Plot distribution of haplotype partitioned reads for more than two
haplotypes}
\usage{
plotPartitionHistogramMulti(x, limits, label = "")
}
\arguments{
\item{x}{A \code{HapPart} object.}

\item{limits}{provided limits for each haplotype as a named list.}

\item{label}{Optional plot label.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot distribution of haplotype partitioned reads for more than two
haplotypes
}
\examples{
###

}
