% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionLongReads.R
\name{plotPartitionTree}
\alias{plotPartitionTree}
\title{Plot tree of initial clustering}
\usage{
plotPartitionTree(x)
}
\arguments{
\item{x}{A \code{HapPart} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot tree of initial clustering
}
\examples{
###
}
