% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup.R
\name{plotPileupCoverage}
\alias{plotPileupCoverage}
\title{Plot pileup coverage}
\usage{
plotPileupCoverage(x, threshold = 0.2, range = NULL, thin = 0.1,
  width = 1, label = "", drop.indels = FALSE,
  compareReference = FALSE)
}
\arguments{
\item{x}{A \code{pileup} object.}

\item{threshold}{At which frequency do we visualise polymorphisms.}

\item{range}{Which part along the x-axis do we plot.}

\item{thin}{Plot only a fraction of the (non-polymorphic) positions.}

\item{width}{Width of polymorphic positions.}

\item{label}{Optional plot label.}

\item{drop.indels}{Don't plot indels.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot pileup coverage
}
\examples{
###
}
