% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-reads.R
\name{plotRadarPartition}
\alias{plotRadarPartition}
\title{Plot a radar chart for each haplotype. Membership values for each read
for each partition are shown as lines.}
\usage{
plotRadarPartition(x)
}
\arguments{
\item{x}{A \code{HapPart} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a radar chart for each haplotype. Membership values for each read
for each partition are shown as lines.
}
\examples{
###
}
