% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymorphic-positions.R
\name{polymorphic_positions}
\alias{polymorphic_positions}
\title{Calculate polymorphic positions from a consensus matrix.}
\usage{
polymorphic_positions(x, theshold, ...)
}
\arguments{
\item{x}{A consensus matrix or pileup object.}

\item{theshold}{Threshold.}

\item{...}{Additional arguments.}
}
\value{
A data.frame
}
\description{
Calculate polymorphic positions from a consensus matrix.
}
\examples{
##
}
