% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{readMSA}
\alias{readMSA}
\title{Read a multiple sequence alignment from file.}
\usage{
readMSA(file)
}
\arguments{
\item{file}{The absolute path to the file storing the MSA.}
}
\description{
Read a multiple sequence alignment from file.
}
\details{
The format  a \code{phylip} like format written by
\code{\link{writeMSA}}.
}
