% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{readpath}
\alias{readpath}
\title{Access the path to a read sequence file.}
\usage{
readpath(x, ...)
}
\arguments{
\item{x}{An object containing paths to read files.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the path to a read sequence file.
}
\examples{
###
}
