% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{readtype}
\alias{readtype}
\title{Access the readtype.}
\usage{
readtype(x, ...)
}
\arguments{
\item{x}{An object containing readtype information.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the readtype.
}
\examples{
###
}
