% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{refname}
\alias{refname}
\title{Access the reference name.}
\usage{
refname(x, ...)
}
\arguments{
\item{x}{An object containing reference name information.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the reference name.
}
\examples{
###
}
