% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{refpath}
\alias{refpath}
\title{Access the reference path of an object}
\usage{
refpath(x, ...)
}
\arguments{
\item{x}{An object containing paths}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the reference path of an object
}
\examples{
###
}
