% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{report}
\alias{report}
\title{Report the final haplotype sequences.}
\usage{
report(x, which, blockWidth = 80, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{which}{Which mapping do we want to report. Will choose \code{mapFinal},
\code{mapIter} in this order if available.}

\item{blockWidth}{Maximum number of sequence letters per line in pairwise
alignment.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Side effect
}
\description{
Report the final haplotype sequences.
}
\details{
\code{report} will create a \code{report} directory containing the consensus
sequences \code{map\{1,2,3\}.\{A,B\}.\{readtype\}.\{mapper\}.unchecked.fa},
an html alignment file
\code{map\{1,2,3\}.aln.\{readtype\}.\{mapper\}.unchecked.html},
a pairwise alignment file
\code{map\{1,2,3\}.\{A,B\}.\{readtype\}.\{mapper\}.unchecked.pair},
and a tab-separated file
\code{problems.\{readtype\}.\{mapper\}.tsv} reporting
positions with potentially problematic results.

To finalise the consensus sequences open the pairwise alignment file and save
it replacing \strong{unchecked} with \strong{checked} in the file name.
In the \strong{checked} file manual edits can be performed. After the
pairwise alignment is verified call \code{\link{reportCheckedConsensus}}.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01"
  ) \%>\%
  mapInit() \%>\%
  partitionLongreads() \%>\%
  mapIter() \%>\%
  partitionShortreads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{createDR2SConf}},
  \code{\link{mapFinal}}, \code{\link{mapInit}},
  \code{\link{mapIter}}, \code{\link{partitionLongreads}},
  \code{\link{polish}}, \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}}
}
\concept{DR2S mapper functions}
