% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{reportCheckedConsensus}
\alias{reportCheckedConsensus}
\title{Report haplotype consensus files as verified}
\usage{
reportCheckedConsensus(x, which = "mapFinal")
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{which}{Which mapping do we want to update.}
}
\value{
Side effect
}
\description{
Report haplotype consensus files as verified
}
\details{
\code{reportCheckedConsensus} will create a \code{checked} directory
containing the consensus sequences
\code{\{A,B\}.\{readtype\}.\{mapper\}.checked.fa}
and a html alignment file \code{aln.\{readtype\}.\{mapper\}.checked.html}.
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{createDR2SConf}},
  \code{\link{mapFinal}}, \code{\link{mapInit}},
  \code{\link{mapIter}}, \code{\link{partitionLongReads}},
  \code{\link{polish}}, \code{\link{refineAlignment}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
