% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionShortReads.R
\name{scoreHighestSR}
\alias{scoreHighestSR}
\title{Get the highest scoring haplotype for each read}
\usage{
scoreHighestSR(srpartition, diffThreshold = 0.001)
}
\arguments{
\item{srpartition}{srpartition dataframe with scores per read and haplotype.}

\item{diffThreshold}{report reads which scores differ only
below threshold to both haplotypes. Iteratively increases diffThreshold until
all positions differ below threshold.}
}
\value{
A \code{data.frame} object. See \code{Details}.
}
\description{
Get the highest scoring haplotype for each read
}
\details{
Returns a \code{data.frame} object with columns:
\describe{
  \item{read}{<character>; The read name}
  \item{Haplotype}{<character>; Haplotype}
}
}
\examples{
### Score
}
