% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{split_reads_by_haplotype}
\alias{split_reads_by_haplotype}
\title{Partition mapped long reads into haplotypes}
\usage{
split_reads_by_haplotype(x, limits, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{limits}{List of Cutoff values for haplotypes.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Partition mapped long reads into haplotypes
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "multialign"
  ) \%>\%
  clear() \%>\%
  map0() \%>\%
  partition_haplotypes() \%>\%
  split_reads_by_haplotype() \%>\%
  extract_fastq() \%>\%
  map1() \%>\%
  map2() \%>\%
  map3() \%>\%
  cache() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extract_fastq}}, \code{\link{map0}},
  \code{\link{map1}}, \code{\link{map2}},
  \code{\link{map3}}, \code{\link{partition_haplotypes}},
  \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}}
}
