% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{stats}
\alias{stats}
\title{Access the stats path of an object}
\usage{
stats(x, ...)
}
\arguments{
\item{x}{An object containing stats}

\item{...}{Additional arguments.}
}
\value{
A named list.
}
\description{
Access the stats path of an object
}
\examples{
###
}
