% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{tag}
\alias{tag}
\title{Access the name tag of an object.}
\usage{
tag(x, ...)
}
\arguments{
\item{x}{An object containing name tags.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the name tag of an object.
}
\examples{
###
}
