% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{writeMSA}
\alias{writeMSA}
\title{Write a multiple sequence alignment in a phylip like format.
Formatted for easily accessing positions and differences}
\usage{
writeMSA(aln, file = "", block.width = 50)
}
\arguments{
\item{aln}{The alignment as a \code{\link[Biostrings]{DNAStringSet}}.}

\item{file}{The filename and path.}

\item{block.width}{The block width of the resulting alignment.}
}
\description{
Write a multiple sequence alignment in a phylip like format.
Formatted for easily accessing positions and differences
}
\examples{
\dontrun{
library(Biostrings)
msa <- DNAStringSet(c("AAA", "AGA", ))
print("TODO: write rep seq example which runs")
}
}
