% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{check_alignment_file}
\alias{check_alignment_file}
\title{Manually check the alignment of consensus sequences using an editor.}
\usage{
check_alignment_file(x, which = "mapFinal", where = 0,
  editor = "xdg-open", openEditor = TRUE)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{which}{Which stage of mapping to use. Either "mapFinal" (default) or
"mapIter".}

\item{where}{Where to focus with the editor.}

\item{editor}{Which editor to use. Either "xdg-open" (default) for standard
system editor, "subl" for sublime, "gvim" or "gedit"}
}
\description{
Manually check the alignment of consensus sequences using an editor.
}
