% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequtils.R
\name{checkHomopolymerCount}
\alias{checkHomopolymerCount}
\title{Get a distribution of homopolymer counts in alleles}
\usage{
checkHomopolymerCount(x, hpCount = 10, map = "mapFinal")
}
\arguments{
\item{x}{a DR2S object.}

\item{hpCount}{the minimal length of a homopolymer to be checked (10).}

\item{map}{Which result to use. Either "mapFinal" or "remap".}
}
\value{
plot a pdf with length histogram and write mode to log
}
\description{
Get a distribution of homopolymer counts in alleles
}
\examples{
###
}
