% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conseq.R
\name{conseq}
\alias{conseq}
\title{Construct a consensus matrix from a pile-up}
\usage{
conseq(x, ...)
}
\arguments{
\item{x}{\code{pileup} object.}

\item{...}{Additional arguments to \code{prune_consensus_matrix}.
Defaults \code{n_look_behind = 12} and \code{cutoff = 0.5}.}

\item{name}{Name for consensus sequence.}

\item{type}{One of "prob", "freq", or "ambig".}

\item{threshold}{If \code{type == "ambig"}, threshold to call an ambiguous
consensus call.}

\item{exclude_gaps}{Exclude gaps at insertion position from consensus
calling.}

\item{prune_matrix}{Exclude positions with a sudden drop in coverage from
consensus calling.}
}
\value{
A \code{\linkS4class{BStringSet}} object with a metadata list
containing the slots:
\describe{
  \item{zscore}{}
  \item{freq}{}
  \item{ambigs}{}
  \item{insertions}{}
  \item{deletions}{}
  \item{consmat}{}
}
}
\description{
Construct a consensus matrix from a pile-up
}
\examples{
###
}
