% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conseq.R
\name{conseq}
\alias{conseq}
\title{Construct a consensus sequence}
\usage{
conseq(x, ...)
}
\arguments{
\item{x}{\code{pileup} object.}

\item{...}{Additional arguments.}

\item{name}{Name for consensus sequence.}

\item{type}{One of "prob" or "ambig".}

\item{threshold}{If \code{type == "ambig"}, threshold to call an ambiguous
consensus call.}

\item{excludeGaps}{Exclude gaps at insertion position from consensus
calling.}

\item{gapSuppressionRatio}{The ratio of base/gap above which gaps at
insertion position are excluded from from consensus calling.}

\item{forceExcludeGaps}{Exclude gaps at from consensus calling irrespective
of the frequency of the gap.}
}
\value{
A \code{\linkS4class{BStringSet}} object with a metadata list
containing the slots:
\describe{
  \item{zscore}{}
  \item{freq}{}
  \item{ambigs}{}
  \item{insertions}{}
  \item{deletions}{}
  \item{consmat}{}
}
}
\description{
Construct a consensus sequence
}
\examples{
###
}
