% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consmat.R
\name{consmat}
\alias{consmat}
\title{Construct a consensus matrix from pileup}
\usage{
consmat(x, ...)
}
\arguments{
\item{x}{A \code{pileup} object.}

\item{...}{Additional arguments such as \code{n}, \code{offset}, 
\code{insertions}}

\item{freq}{If \code{TRUE} then frequencies are reported, otherwise counts.}
}
\value{
A \code{consmat} object.
}
\description{
Construct a consensus matrix from pileup
}
\details{
\code{consmat}: a \code{matrix} with positions row names and nucleotides as
column manes.
A \code{consmat} object includes the attributes:
\describe{
  \item{n}{<integer>; Number of reads per row.}
  \item{freq}{<logical>; Is it a frequency matrix or a count matrix.}
  \item{offset}{<integer>; Offset}
  \item{insertions}{<integer>; Insertions}
}
}
\examples{
###
}
