% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{createDR2SConf}
\alias{createDR2SConf}
\title{Create a \code{\link[=DR2S_]{DR2S}} configuration object.}
\usage{
createDR2SConf(sample, locus, longreads = list(dir = "pacbio", type =
  "pacbio", mapper = "minimap"), shortreads = NULL, datadir = ".",
  outdir = "./output", reference = NULL, details = NULL,
  opts = NULL, ...)
}
\arguments{
\item{sample}{A unique sample identifier used to locate the long and short
read FASTQ/FASTA files.}

\item{locus}{The HLA or KIR locus (e.g., "A", "DPB1", or "2DL1"; see Note).}

\item{longreads}{Location, type, and mapper for longreads as a named list
with the fields \code{dir}, \code{type} ("pacbio" or "nanopore") and
\code{mapper} ("bwamem" or "minimap").}

\item{shortreads}{(optional) Location, type, and mapper for short reads
as a named list with the fields \code{dir}, \code{type} ("illumina") and
\code{mapper} ("bwamem" or "minimap").}

\item{datadir}{The data directory (See Note).}

\item{outdir}{The output directory (See Note).}

\item{reference}{(optional) Path to reference sequence (See Note).}

\item{details}{<named list> of sample metadata or \code{NULL}. Will be written
into the fasta header of the final sequences and stored in the config json}

\item{opts}{<named list> of arguments to the DR2S pipeline functions or
\code{NULL}. Will be stored in the config json.}

\item{...}{Further arguments.}
}
\value{
A \code{\link[DR2S]{DR2S}} config object
}
\description{
Create a \code{\link[=DR2S_]{DR2S}} configuration object.
}
\section{locus}{

The locus can be specified generically, e.g. "A", "B", or "2DL1". In this
case a generic, locus-specific reference will initially be used.
Alternatively a locus can be specified more specifically by including
allele code information (if available), e.g. "DPB1*04:01:01:01". In this
case the specified allele will be used as initial reference.
}

\section{datadir}{

A \code{datadir} must contain subdirectories (default: "pacbio" and
"illumina") containig long read FASTQs/FASTAs in the format
\code{SAMPLE_LOCUS_*.fast(q|a)(.gz)} and short read FASTQs/FASTAs in the
format \code{SAMPLE_LOCUS_*R[12]*.fast(q|a)(.gz)}, respectively.
}

\section{Outdir}{

All output will be placed in directory hierarchy \code{OUTDIR/SAMPLE/LOCUS}
}

\section{Reference}{

References can be specified as a path to a fasta file containing the
reference sequence.
}

\examples{
\dontrun{
x <- InitDR2S(createDR2SConf(
  sample = "ID12300527",
  locus = "DPB1*04:01:01:01",
  datadir = "/path/to/data",
  outdir = "/path/to/output"
  )) \%>\%
  mapInit() \%>\%
  partitionLongreads() \%>\%
  mapIter() \%>\%
  partitionShortreads() \%>\%
  mapFinal() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{mapFinal}},
  \code{\link{mapInit}}, \code{\link{mapIter}},
  \code{\link{partitionLongreads}},
  \code{\link{remapAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
