% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{createDR2SConf}
\alias{createDR2SConf}
\title{Create a \code{\link[=DR2S_]{DR2S}} configuration object.}
\usage{
createDR2SConf(sample, locus, longreads = list(dir = "pacbio", type =
  "pacbio", mapper = "minimap"), shortreads = NULL, datadir = ".",
  outdir = "./output", reference = NULL, threshold = 0.2,
  iterations = 1, microsatellite = FALSE, distAlleles = 2,
  filterScores = FALSE, forceMapping = FALSE, details = list(), ...)
}
\arguments{
\item{sample}{A unique sample identifier used to locate the long and short
read FASTQ/FASTA files.}

\item{locus}{The HLA or KIR locus (e.g., "A", "DPB1", or "2DL1")}

\item{longreads}{Location, type, and mapper for longreads as a named list
with the fields \code{dir}, \code{type} ("pacbio" or "nanopore") and
\code{mapper} ("bwamem" or "minimap").}

\item{shortreads}{(optional) Location, type, and mapper for short reads
as a named list with the fields \code{dir}, \code{type} ("illumina") and
\code{mapper} ("bwamem" or "minimap").}

\item{datadir}{The data directory (See Note).}

\item{outdir}{The output directory (See Note).}

\item{reference}{The reference allele (See Note).}

\item{threshold}{Threshold frequency for detecting polymorphisms.}

\item{iterations}{Number of iterations of the mapIter step.}

\item{microsatellite}{<\code{FALSE}>; Perform a second mapping of shortreads
to the inferred reference in mapInit. Set to TRUE if you suspect
microsatellites or other repetitive regions in your sequence. Usually extends
the reference to a maximum length and enables a better mapping.}

\item{distAlleles}{Number of different alleles in the sample. Should be 2
for heterozygous samples, 1 for homozygous samples and > 2 for some KIR loci.}

\item{filterScores}{use only reads passing a strict filtering step. TODO}

\item{forceMapping}{<\code{FALSE}>; set to TRUE if you want to force
processing of "bad" shortreads, i.e. when the distribution of coverage is
heavily unequal. Aborts the program if maximum coverage > 75 \% quantile * 5.}

\item{details}{Named list of sample metadata or \code{NULL}. Will be written
into the fasta header of the final sequences and stored in the config yaml.}

\item{...}{Additional arguments.}
}
\value{
A \code{\link[DR2S]{DR2S}} config object
}
\description{
Create a \code{\link[=DR2S_]{DR2S}} configuration object.
}
\section{datadir}{

A \code{datadir} must contain subdirectories (default: "pacbio" and
"illumina") containig long read FASTQs/FASTAs in the format
\code{SAMPLE_LOCUS_*.fast(q|a)(.gz)} and short read FASTQs/FASTAs in the
format \code{SAMPLE_LOCUS_*R[12]*.fast(q|a)(.gz)}, respectively.
}

\section{Outdir}{

All output will be placed in directory hierarchy \code{OUTDIR/SAMPLE/LOCUS}
}

\section{Reference}{

References can be specified as allele codes or a path to a fasta file
containing the reference sequence.
}

\examples{
\dontrun{
x <- InitDR2S(createDR2SConf(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01"
  )) \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{mapFinal}},
  \code{\link{mapInit}}, \code{\link{mapIter}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
