% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callVariants.R
\name{disambiguateVariant}
\alias{disambiguateVariant}
\title{Disambiguate a variant}
\usage{
disambiguateVariant(x, threshold)
}
\arguments{
\item{x}{A \code{variantList}.}

\item{threshold}{Minimum frequency for a variant.}
}
\value{
A \code{variantList}.
}
\description{
Disambiguate a variant
}
\examples{
###

}
\keyword{internal}
