% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequtils.R
\name{.generateReferenceSequence}
\alias{.generateReferenceSequence}
\title{Generate reference sequences in a FASTA file}
\usage{
.generateReferenceSequence(allele, locus, outdir, dirtag = NULL)
}
\arguments{
\item{allele}{Allele name.}

\item{outdir}{Output directory}

\item{dirtag}{TODO}

\item{HLA}{A \code{HLAGene} object}
}
\value{
Path to refseq fasta file.
}
\description{
Generate reference sequences in a FASTA file
}
\examples{
###
}
\keyword{internal}
