% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getIdleCores}
\alias{.getIdleCores}
\title{Get the number of idling cores that can be used by calling the linux system
program mpstat}
\usage{
.getIdleCores()
}
\value{
An integer giving the number of idling cores
}
\description{
Get the number of idling cores that can be used by calling the linux system
program mpstat
}
