% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR2S-class.R
\docType{class}
\name{DR2S_}
\alias{DR2S_}
\title{Class \code{"DR2S"}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DR2Smap(sample, locus, longreads = list(type = "pacbio", dir = "pacbio"),
shortreads = list(type = "illumina", dir = "illumina"), datadir = ".",
outdir = "./output", reference = NULL, consensus = "multialign",
threshold = 0.20, fullname = TRUE, create_outdir = TRUE, ...)
}
\value{
Object of \code{\link{R6Class}} representing a DR2S analysis.
}
\description{
Class \code{"DR2S"}
}
\section{Fields}{

\describe{
\item{\code{map0}}{\code{[map0]}; the mapping of long reads to \code{reference}.}

\item{\code{partition}}{\code{[PartList]}; the partitioning of full-length mapped
long reads into two haplotypes A and B.}

\item{\code{map1}}{\code{[map1]}; the mapping of A and B reads to reference sequences
generated by multiple alignment of a subset of A and B reads.}

\item{\code{map2}}{\code{[map2]}; the mapping of A and B reads to consensus
sequences produced in the previous step.}

\item{\code{map3}}{\code{[map3]}; the mapping of A, B, and short reads to consensus
sequences produced in the previous step.}

\item{\code{consensus}}{\code{[ConsList]}; final consensus sequences for A and B.}
}}

\section{Public Methods}{

\describe{
\item{\code{x$runMap0(opts = list(), optsname = "", pct = 100, threshold = 0.20,
min_base_quality = 3, min_mapq = 0, max_depth = 1e4, min_nucleotide_depth = 3,
include_deletions = FALSE, include_insertions = FALSE, force = FALSE,
fullname = TRUE, plot = TRUE)}}{
Run the inital mapping step (long reads against the reference allele)}
\item{\code{x$runHaplotypePartitioning(max_depth = 1e4, shuffle = TRUE,
skip_gap_freq = 2/3, plot = TRUE)}}{
Partition mapped long reads into haplotypes}
\item{\code{x$splitReadsByHaplotype(limit = list(), #' plot = TRUE)}}{
Partition mapped long reads into haplotypes}
\item{\code{x$extractFastq(nreads = NULL, replace = FALSE, nalign = 30)}}{
Extract FASTQs for partitioned reads and if the consensus method is
\code{multialign} contruct a multiple sequence alignment from the best
\code{nalign} reads.}
}
}

\section{Internals}{

\describe{
\item{\code{x$new(conf, create_outdir = TRUE)}}{Initialise object of class
 \code{DR2S}.}
\item{x$getConfig(name = NULL}{Extract configuration}
\item{x$setConfig(name, value}{Modify configuration}
}
}

\keyword{data}
\keyword{internal}
