% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR2S-class.R
\docType{class}
\name{DR2S_}
\alias{DR2S_}
\title{Class \code{"DR2S"}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DR2Smap(sample, locus, longreads = list(type = "pacbio", 
dir = "pacbio"), shortreads = list(type = "illumina", dir = "illumina"), 
datadir = ".", outdir = "./output", reference = NULL,,
threshold = 0.20, iterations = 1, microsatellite = FALSE, distAlleles = 2,
filterScores = TRUE, partSR = TRUE, fullname = TRUE, createOutdir = TRUE, 
...)
}
\value{
Object of \code{\link{R6Class}} representing a DR2S analysis.
}
\description{
Class \code{"DR2S"}
}
\section{Fields}{

\describe{
\item{\code{mapInit}}{\code{[mapInit]}; the mapping of long reads to 
\code{reference}.}

\item{\code{partition}}{\code{[PartList]}; the partitioning of full-length mapped
long reads into different haplotypes.}

\item{\code{mapIter}}{\code{[mapIter]}; the mapping of A and B reads to consensus
sequences produced in the previous step.}

\item{\code{mapFinal}}{\code{[mapFinal]}; the mapping of A, B, and short reads to 
consensus sequences produced in the previous step.}

\item{\code{consensus}}{\code{[ConsList]}; final consensus sequences for A and B.}
}}

\section{Public Methods}{

\describe{
\item{\code{x$runMapInit(opts = list(), optsname = "", pct = 100, 
threshold = 0.20, iterations = 1, microsatellite = FALSE, distAlleles = 2, 
filterScores = TRUE, partSR = TRUE, minBaseQuality = 3, minMapq = 0, 
maxDepth = 1e4, minNucleotideDepth = 3, includeDeletions = FALSE, 
includeInsertions = FALSE, force = FALSE, fullname = TRUE, plot = TRUE)}}{
Run the inital mapping step (long reads against the reference allele)}
\item{\code{x$runHaplotypePartitioning(maxDepth = 1e4, shuffle = TRUE,
skipGapFreq = 2/3, plot = TRUE)}}{
Partition mapped long reads into haplotypes}
\item{\code{x$splitReadsByHaplotype(limit = list(), #' plot = TRUE)}}{
Partition mapped long reads into haplotypes}
\item{\code{x$extractFastq()}}{
Extract FASTQs for partitioned reads.
}}
}

\section{Internals}{

\describe{
\item{\code{x$new(conf, createOutdir = TRUE)}}{Initialise object of class
 \code{DR2S}.}
\item{x$getConfig(name = NULL}{Extract configuration}
\item{x$setConfig(name, value}{Modify configuration}
}
}

\keyword{data}
\keyword{internal}
