% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR2S-class.R
\docType{class}
\name{DR2S_}
\alias{DR2S_}
\title{Class \code{"DR2S"}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InitDR2S(config, createOutdir = TRUE)
}
\value{
Object of \code{\link[R6]{R6Class}} representing a DR2S analysis.
}
\description{
Class \code{"DR2S"}
}
\section{Fields}{

\describe{
\item{\code{InitDR2S}}{Initialize DR2S from a config.}

\item{\code{mapInit}}{\code{[MapList]}; the mapping of long reads to
\code{reference}.}

\item{\code{partition}}{\code{[PartList]}; the partitioning of full-length
mapped long reads into different haplotypes.}

\item{\code{mapIter}}{\code{[MapList]}; the mapping of A and B reads to
consensus sequences produced in the previous step.}

\item{\code{mapFinal}}{\code{[MapList]}; the mapping of A, B, and short reads
to consensus sequences produced in the previous step.}

\item{\code{consensus}}{\code{[ConsList]}; final consensus sequences for A and
B.}
}}

\section{Public Methods}{

\describe{
\item{\code{x$runMapInit(opts = list(), optsname = "",  threshold = 0.20,
iterations = 1, microsatellite = FALSE, distAlleles = 2, filterScores = TRUE,
minBaseQuality = 3, minMapq = 0, maxDepth = 1e4, minNucleotideDepth = 3,
includeDeletions = FALSE, includeInsertions = FALSE,
force = FALSE, plot = TRUE)}}{Run the inital mapping step
(long reads against the reference allele)}
\item{\code{x$runHaplotypePartitioning(maxDepth = 1e4, shuffle = TRUE,
skipGapFreq = 2/3, plot = TRUE)}}{
Partition mapped long reads into haplotypes}
\item{\code{x$splitReadsByHaplotype(limit = list(), #' plot = TRUE)}}{
Partition mapped long reads into haplotypes}
\item{\code{x$extractFastq()}}{
Extract FASTQs for partitioned reads.
}}
}

\section{Internals}{

\describe{
\item{\code{x$new(conf, createOutdir = TRUE)}}{Initialise object of class
 \code{DR2S}.}
\item{x$getConfig(name = NULL}{Extract configuration}
\item{x$setConfig(name, value}{Modify configuration}
}
}

\keyword{data}
\keyword{internal}
