% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{DR2Smap}
\alias{DR2Smap}
\title{Constructor for \code{\link[=DR2S_]{DR2S}} mapper objects.}
\usage{
DR2Smap(sample, locus, longreads = list(type = "pacbio", dir = "pacbio"),
  shortreads = list(type = "illumina", dir = "illumina"), datadir = ".",
  outdir = "./output", reference = NULL, consensus = "mapping",
  threshold = 0.2, iterations = 1, partSR = TRUE, fullname = TRUE, ...)
}
\arguments{
\item{sample}{A unique sample identifier used to locate the long and short
read FASTQ files.}

\item{locus}{The HLA locus ("A", "B", "C", "DPB1", "DQB1", "DRB1") or KIR locus.}

\item{longreads}{The type and location of the long reads as a named list
with the fields \code{type} ("pacbio" or "nanopore") and \code{dir}.
Additional optional fields: \code{name} and \code{opts}.}

\item{shortreads}{The type and location of the short reads as a named list
with the fields \code{type} ("illumina") and \code{dir}. Short reads are optional.
\code{NULL} can be provided.}

\item{datadir}{The data directory (See Note).}

\item{outdir}{The output directory (See Note).}

\item{reference}{The reference allele(s).}

\item{consensus}{\dQuote{\code{mapping}} for now (See Note).}

\item{threshold}{Threshold frequency for polymorphisms.}

\item{iterations}{Number of iterations of the mapIter step.}

\item{partSR}{Use shortreads in the mapInit step for getting polymorphic positions and
a first reference.}

\item{fullname}{Truncate allele names.}

\item{microsatellite}{FALSE Perform a second mapping of shortreads to the inferred reference in
mapInit. Set to TRUE if you know you have repeats like in microsatellites. Usually extends the
reference to a maximum length and enables a better mapping.}

\item{forceBadMapping}{FALSE set to TRUE if you want to force processing of bad shortreads, i.e.
when the distribution of coverage is bad. Aborts the program if maximum coverage > 75 % quantile * 5.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Initialise a \code{\link[=DR2S_]{DR2S}} mapper.
}
\section{datadir}{


A \code{datadir} must contain arbitrarily named subdirectories
(default: "pacbio" and "illumina") with long read FASTQs in the format
\code{SAMPLE_LOCUS_*.fastq(.gz)} and short read FASTQs in the format
\code{SAMPLE_LOCUS_*R[12]*.fastq(.gz)}, respectively.
}

\section{Outdir}{


All output will be placed in directory hierarchy
\code{OUTDIR/SAMPLE/READTYPE/LOCUS}.
}

\section{Reference}{


References can be specified as allele codes or a path to a fasta file
containing the reference sequence. If \code{reference = NULL} a global generic reference
for a given locus will be used.
}

\section{Consensus}{

Right now, the consensus method of multialign is removed. A de novo assembly of
longreads, maybe supported by shortreads might be implemented.
\dQuote{\bold{\code{mapping}}}: The reference is refined from the initial provided reference
during the mapInit step using short reads. Individual references for each found haplotype are
constructed from this reference using only longreads that are assigned to the haplotype.
}

\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  ) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partition_haplotypes() \%>\%
  split_reads_by_haplotype() \%>\%
  extract_fastq() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{extract_fastq}},
  \code{\link{mapFinal}}, \code{\link{mapInit}},
  \code{\link{mapIter}}, \code{\link{partitionShortReads}},
  \code{\link{partition_haplotypes}}, \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}},
  \code{\link{split_reads_by_haplotype}}
}
