% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{InitDR2S}
\alias{InitDR2S}
\title{Initialise a DR2S instance}
\usage{
InitDR2S(config, createOutdir = TRUE)
}
\arguments{
\item{config}{A DR2S configuration object.}

\item{createOutdir}{Create the outdir if not exists.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Initialise a DR2S instance
}
