% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapFinal}
\alias{mapFinal}
\title{Final mapping (long reads and short reads).}
\usage{
mapFinal(x, opts = list(), includeDeletions = TRUE,
  includeInsertions = TRUE, clip = TRUE, force = FALSE,
  createIgv = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{includeDeletions}{If \code{TRUE}, include deletions in pileup.}

\item{includeInsertions}{If \code{TRUE}, include insertions in pileup.
from bam files for
 variants ar polymorphic positions.}

\item{clip}{Clip homopolymeric read ends from shortreads.}

\item{force}{If \code{TRUE}, overwrite existing bam file.}

\item{createIgv}{Subsample for looking with IgvJs in the shiny app.}

\item{plot}{Plot diagnostics.}

\item{...}{Additional parameters passed to \code{\link[Rsamtools]{PileupParam}}.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Partitioned long reads and partitioned or unpartitioned short reads are
mapped against the consensus sequences inferred at \code{mapIter}. Indels
are \strong{included} in pileup.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01"
  ) \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{createDR2SConf}},
  \code{\link{mapInit}}, \code{\link{mapIter}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
