% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapFinal}
\alias{mapFinal}
\title{Final mapping step (long reads and short reads).}
\usage{
mapFinal(x, opts = list(), pct = 100, min_base_quality = 7,
  min_mapq = 0, max_depth = 1e+05, min_nucleotide_depth = 3,
  include_deletions = TRUE, include_insertions = TRUE,
  include_read_ids = TRUE, force = FALSE, fullname = TRUE, plot = TRUE,
  clip = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{pct}{Percentage of templates to subsample (before sorting the bam file).}

\item{min_base_quality}{Minimum \sQuote{QUAL} value for each nucleotide in an
alignment.}

\item{min_mapq}{Minimum \sQuote{MAPQ} value for an alignment to be included
in pileup.}

\item{max_depth}{Maximum number of overlapping alignments considered for
each position in the pileup.}

\item{min_nucleotide_depth}{Minimum count of each nucleotide at a given
position required for that nucleotide to appear in the result.}

\item{include_deletions}{If \code{TRUE}, include deletions in pileup.}

\item{include_insertions}{If \code{TRUE}, include insertions in pileup.}

\item{include_read_ids}{If \code{TRUE}, read identifieres are extracted
from bam files for variants ar polymorphic positions.}

\item{force}{If \code{TRUE}, overwrite existing bam file.}

\item{fullname}{If \code{TRUE}, use the full name in reference fasta.}

\item{plot}{Plot diagnostics.}

\item{clip}{Clip homopolymeric read ends from shortreads.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Partitioned long reads and partitioned or unpartitioned short reads are mapped against the
consensus sequences inferred at \code{mapIter}. Indels are \strong{included}
in pileup.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  ) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partition_haplotypes() \%>\%
  split_reads_by_haplotype() \%>\%
  extract_fastq() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extract_fastq}}, \code{\link{mapInit}},
  \code{\link{mapIter}},
  \code{\link{partition_haplotypes}}, \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}},
  \code{\link{split_reads_by_haplotype}}
}
