% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapFinal}
\alias{mapFinal}
\title{Final mapping step (long reads and short reads).}
\usage{
mapFinal(x, opts = list(), pct = 100, minBaseQuality = 7, minMapq = 0,
  maxDepth = 1e+05, minNucleotideDepth = 3, includeDeletions = TRUE,
  includeInsertions = TRUE, force = FALSE, fullname = TRUE, plot = TRUE,
  clip = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{pct}{Percentage of templates to subsample (before sorting the bam file).}

\item{minBaseQuality}{Minimum \sQuote{QUAL} value for each nucleotide in an
alignment.}

\item{minMapq}{Minimum \sQuote{MAPQ} value for an alignment to be included
in pileup.}

\item{maxDepth}{Maximum number of overlapping alignments considered for
each position in the pileup.}

\item{minNucleotideDepth}{Minimum count of each nucleotide at a given
position required for that nucleotide to appear in the result.}

\item{includeDeletions}{If \code{TRUE}, include deletions in pileup.}

\item{includeInsertions}{If \code{TRUE}, include insertions in pileup.
from bam files for variants ar polymorphic positions.}

\item{force}{If \code{TRUE}, overwrite existing bam file.}

\item{fullname}{If \code{TRUE}, use the full name in reference fasta.}

\item{plot}{Plot diagnostics.}

\item{clip}{Clip homopolymeric read ends from shortreads.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Partitioned long reads and partitioned or unpartitioned short reads are mapped against the
consensus sequences inferred at \code{mapIter}. Indels are \strong{included}
in pileup.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  ) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{mapInit}}, \code{\link{mapIter}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
