% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapInit}
\alias{mapInit}
\title{Initial mapping.}
\usage{
mapInit(x, opts = list(), createIgv = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{List with options passed to the mapper.}

\item{createIgv}{Subsample bam files for visualisation with IgvJs in the
DR2S shiny app.}

\item{plot}{Generate diagnostic plots.}

\item{...}{Additional parameters passed to \code{\link[Rsamtools]{PileupParam}}.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Short reads (Illumina) and long reads (PacBio or Nanopore) or long reads alone
are mapped to a single reference sequence.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01"
  ) \%>\%
  mapInit() \%>\%
  partitionLongreads() \%>\%
  mapIter() \%>\%
  partitionShortreads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{createDR2SConf}},
  \code{\link{mapFinal}}, \code{\link{mapIter}},
  \code{\link{partitionLongreads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
