% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapInit}
\alias{mapInit}
\title{Initial mapping step (long reads).}
\usage{
mapInit(x, opts = list(), refname = "ref", optsname = "",
  threshold = 0.2, minBaseQuality = 7, minMapq = 0, maxDepth = 10000,
  minNucleotideDepth = 3, includeDeletions = FALSE,
  includeInsertions = FALSE, force = FALSE, fullname = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{optsname}{Additional text to describe the options used.}

\item{threshold}{Threshold to call a variant.}

\item{minBaseQuality}{Minimum \sQuote{QUAL} value for each nucleotide in an
alignment.}

\item{minMapq}{Minimum \sQuote{MAPQ} value for an alignment to be included
in pileup.}

\item{maxDepth}{Maximum number of overlapping alignments considered for
each position in the pileup.}

\item{minNucleotideDepth}{Minimum count of each nucleotide at a given
position required for that nucleotide to appear in the result.}

\item{includeDeletions}{If \code{TRUE}, include deletions in pileup.}

\item{includeInsertions}{If \code{TRUE}, include insertions in pileup.}

\item{force}{If \code{TRUE}, overwrite existing bam file.}

\item{fullname}{If \code{TRUE}, use the full name in reference fasta.}

\item{plot}{Produce diagnostic plots.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Long reads (PacBio or Nanopore) of a heterozygous sample are mapped to a
single reference sequence.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  ) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{mapFinal}}, \code{\link{mapIter}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
