% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapInit}
\alias{mapInit}
\title{Initial mapping.}
\usage{
mapInit(x, opts = list(), threshold = NULL, includeDeletions = TRUE,
  includeInsertions = TRUE, microsatellite = FALSE,
  filterScores = TRUE, forceMapping = FALSE, topx = 0,
  createIgv = TRUE, force = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{threshold}{Threshold to call a variant.}

\item{includeDeletions}{If \code{TRUE}, include deletions in pileup.}

\item{includeInsertions}{If \code{TRUE}, include insertions in pileup.}

\item{microsatellite}{If \code{TRUE} remap the shortreads again to expand
the reference to a maximum length. Works better for larger insertions and
repeats like microsatellites.}

\item{filterScores}{Apply a harsh filtering on the reads. Filters for mapping
quality and removes all softclipping reads. Usually not necessary.}

\item{forceMapping}{if \code{FALSE} the program throws an error in case the
coverage of shortreads is too different at different parts of the sequence.}

\item{topx}{Select the x best-scoring reads.}

\item{createIgv}{Subsample for looking with IgvJs in the shiny app.}

\item{force}{If \code{TRUE}, overwrite existing bam files.}

\item{plot}{Produce diagnostic plots.}

\item{...}{Additional parameters passed to \code{\link[Rsamtools]{PileupParam}}.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Short reads (Illumina) and long reads (PacBio or Nanopore) or long reads alone
are mapped to a single reference sequence.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01"
  ) \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockWidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{createDR2SConf}},
  \code{\link{mapFinal}}, \code{\link{mapIter}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
