% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapInit}
\alias{mapInit}
\title{Initial mapping step (long reads).}
\usage{
mapInit(x, opts = list(), refname = "ref", optsname = "", pct = 100,
  threshold = 0.2, min_base_quality = 7, min_mapq = 0,
  max_depth = 10000, min_nucleotide_depth = 3, include_deletions = FALSE,
  include_insertions = FALSE, force = FALSE, fullname = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{refname}{Which allele do we use as primary reference; the
\strong{ref}erence allele or the \strong{alt}ernate allele.}

\item{optsname}{Additional text to describe the options used.}

\item{pct}{Percentage of templates to subsample (before sorting the bam file).}

\item{threshold}{Threshold to call a variant.}

\item{min_base_quality}{Minimum \sQuote{QUAL} value for each nucleotide in an
alignment.}

\item{min_mapq}{Minimum \sQuote{MAPQ} value for an alignment to be included
in pileup.}

\item{max_depth}{Maximum number of overlapping alignments considered for
each position in the pileup.}

\item{min_nucleotide_depth}{Minimum count of each nucleotide at a given
position required for that nucleotide to appear in the result.}

\item{include_deletions}{If \code{TRUE}, include deletions in pileup.}

\item{include_insertions}{If \code{TRUE}, include insertions in pileup.}

\item{force}{If \code{TRUE}, overwrite existing bam file.}

\item{fullname}{If \code{TRUE}, use the full name in reference fasta.}

\item{plot}{Produce diagnostic plots.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Long reads (PacBio or Nanopore) of a heterozygous sample are mapped to a
single reference sequence.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  ) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  splitReadsByHaplotype() \%>\%
  extractFastq() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extractLongReads}}, \code{\link{mapFinal}},
  \code{\link{mapIter}}, \code{\link{partitionLongReads}},
  \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}}, \code{\link{splitReadsByHaplotype}}
}
