% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapIter}
\alias{mapIter}
\title{Iterative mapping.}
\usage{
mapIter(x, opts = list(), iterations = 1, columnOccupancy = 0.4,
  force = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{iterations}{Number of \code{mapIter} iterations. How often are the
clustered reads remapped to updated reference sequences.}

\item{columnOccupancy}{Minimum occupancy (1 - fraction of gap) below which
bases at insertion position are excluded from from consensus calling.}

\item{force}{If \code{TRUE}, overwrite existing bam file.}

\item{plot}{Plot diagnostics.}

\item{...}{Additional parameters passed to \code{\link[Rsamtools]{PileupParam}}.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Partitioned long reads are mapped against the MSA consensus sequence of
each haplotype. Indels are \strong{excluded} from pileup.
New reference consensus sequences for both alleles are inferred.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01"
  ) \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockwidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{createDR2SConf}},
  \code{\link{mapFinal}}, \code{\link{mapInit}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
