% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{mapIter}
\alias{mapIter}
\title{Iterative mapping.}
\usage{
mapIter(x, opts = list(), iterations = 1, minBaseQuality = 3,
  minMapq = 0, maxDepth = 10000, minNucleotideDepth = 3,
  gapSuppressionRatio = 2/5, force = FALSE, plot = TRUE)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{opts}{Mapper options.}

\item{iterations}{Number of iterations. How often are the clustered reads
remapped to the updated reference.}

\item{minBaseQuality}{Minimum \sQuote{QUAL} value for each nucleotide in an
alignment.}

\item{minMapq}{Minimum \sQuote{MAPQ} value for an alignment to be included
in pileup.}

\item{maxDepth}{Maximum number of overlapping alignments considered for
each position in the pileup.}

\item{minNucleotideDepth}{Minimum count of each nucleotide at a given
position required for that nucleotide to appear in the result.}

\item{gapSuppressionRatio}{The ratio of base/gap above which gaps at
insertion position are excluded from from consensus calling.}

\item{force}{If \code{TRUE}, overwrite existing bam file.}

\item{plot}{Plot diagnostics.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object.
}
\description{
Partitioned long reads are mapped against the MSA consensus sequence of
the best 40 sequences for each haplotype. Indels are \strong{excluded} from
pileup. New consensus sequences for both alleles are inferred.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01"
  ) \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockwidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{createDR2SConf}},
  \code{\link{mapFinal}}, \code{\link{mapInit}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
