% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-reads.R
\name{partition_reads}
\alias{partition_reads}
\title{Partition a SNP matrix into two haplotype groups}
\usage{
partition_reads(x, shuffle = TRUE, skip_gap_freq = 2/3)
}
\arguments{
\item{x}{A \code{[read x position]} SNP matrix.}

\item{shuffle}{Shuffle positions.}

\item{skip_gap_freq}{Skip a badly behaved polymorphic position if the
frequency of gaps within a haplotype group exceeds \code{skip_gap_freq}.}
}
\value{
A \code{HapPart} object.
}
\description{
Partition a SNP matrix into two haplotype groups
}
\examples{
###
}

