% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/pileup.R
\name{reads}
\alias{reads}
\alias{reads<-}
\alias{indelRate}
\alias{indelRate<-}
\alias{pileup}
\title{Calculate pileup for a BAM file.}
\usage{
reads(x, ...)

reads(x) <- value

indelRate(x, ...)

indelRate(x) <- value

pileup(bamfile, reffile, readtype, ..., pParam)
}
\arguments{
\item{...}{Additional parameters passed to \code{\link[Rsamtools]{PileupParam}}.}

\item{bamfile}{<\code{character}>; path to bamfile.}

\item{reffile}{<\code{character}>; Path to reference.}

\item{readtype}{<\code{character}>; one of "illumina", "pacbio", or "nanopore".}

\item{pParam}{A \code{\link[Rsamtools]{PileupParam}} object.}
}
\value{
A \code{pileup} object. See \code{Details}.
}
\description{
Calculate pileup for a BAM file.
}
\details{
Returns a \code{pileup} object:
A \code{list} with slots:
\describe{
  \item{bamfile}{<character>; Path to the bamfile used to construct the
  pileup}
  \item{reffile}{<character>; Path to the reference used to construct the
  bamfile}
  \item{refname}{<character>; Name of the reference}
  \item{readtype}{<character>; One of "illumina", "pacbio", or "nanopore"}
  \item{param}{A \code{\link[Rsamtools]{PileupParam}} object}
  \item{pilup}{A \code{tbl_df} with colums:
    \describe{
      \item{seqnames}{<character>; The RNAME field}
      \item{pos}{<integer>; Genomic position of base}
      \item{nucleotide}{<character>; Base}
      \item{count}{<integer>; Count of base}
    }
  }
  \item{consmat}{<consmat>; The consensus matrix}
  \item{reads}{<character>; (optional) names of n top-scoring reads}
  \item{stats}{<list>; Run statistics}
}
}
\examples{
###
}
