% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{polish}
\alias{polish}
\title{Polish final haplotype sequences.}
\usage{
polish(x, threshold = x$getThreshold(), lower_limit = 0.6, cache = TRUE,
  ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{threshold}{When do we call a variant a variant.}

\item{lower_limit}{When do we report a variant as uncertain.}

\item{cache}{Cache the updated \code{DR2S} object after assembling the haplotypes.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object with the \code{consensus} field
populated.
}
\description{
Polish final haplotype sequences.
}
\examples{
## Run the full DR2S pipeline
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "multialign"
  ) \%>\%
  clear() \%>\%
  map0() \%>\%
  partition_haplotypes() \%>\%
  split_reads_by_haplotype(limitA = 0.5, limitB = 0.25) \%>\%
  extract_fastq() \%>\%
  map1() \%>\%
  map2() \%>\%
  map3() \%>\%
  cache() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S.mapper.functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extract_fastq}}, \code{\link{map0}},
  \code{\link{map1}}, \code{\link{map2}},
  \code{\link{map3}}, \code{\link{partition_haplotypes}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}},
  \code{\link{split_reads_by_haplotype}}
}

