% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{polish}
\alias{polish}
\title{Polish final haplotype sequences.}
\usage{
polish(x, threshold = x$getThreshold(), checkHpCount = TRUE,
  cache = TRUE)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{threshold}{When do we call a variant a variant.}

\item{checkHpCount}{Check the number of homopolymer counts in shortreads.
Compare the resulting sequence with the mode value and report differences.}

\item{cache}{Cache the updated \code{DR2S} object after assembling the 
haplotypes.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{\link[=DR2S_]{DR2S}} object with the \code{consensus} field
populated.
}
\description{
Polish final haplotype sequences.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  ) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(blockwidth = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{cache}},
  \code{\link{clear}}, \code{\link{createDR2SConf}},
  \code{\link{mapFinal}}, \code{\link{mapInit}},
  \code{\link{mapIter}}, \code{\link{partitionLongReads}},
  \code{\link{refineAlignment}},
  \code{\link{reportCheckedConsensus}},
  \code{\link{report}}
}
\concept{DR2S mapper functions}
