% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{readDR2SConf}
\alias{readDR2SConf}
\title{Read a DR2S config file in yaml format}
\usage{
readDR2SConf(configFile)
}
\arguments{
\item{configFile}{The path to the config file.}
}
\description{
Read a DR2S config file in yaml format
}
\details{
DR2S config files can be created manually or by the 
\code{\link{writeDR2SConf}} function.
}
