% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{refineAlignment}
\alias{refineAlignment}
\title{Refine the mapping of an allele by remap to the checked consensus}
\usage{
refineAlignment(x, hptype)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{hptype}{The allele to refine}
}
\value{
Returns the updated \code{\link[=DR2s_]{DR2S}} object.

Creates an executable bash file for inspecting the mapping with IGV
}
\description{
Refine the mapping of an allele by remap to the checked consensus
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{mapFinal}}, \code{\link{mapInit}},
  \code{\link{mapIter}}, \code{\link{partitionLongReads}},
  \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{report}}
}
