% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{refname}
\alias{refname}
\title{Access the reference name of an object}
\usage{
refname(x, ...)
}
\arguments{
\item{x}{An object containing paths.}

\item{...}{Additional arguments.}
}
\value{
A character vector.
}
\description{
Access the reference name of an object
}
\examples{
###
}
