% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{report}
\alias{report}
\title{Report the final haplotype sequences.}
\usage{
report(x, which, block_width = 80, ...)
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{which}{Which mapping do we want to report. Will choose \code{mapFinal},
\code{mapIter} in this order if available.}

\item{block_width}{Maximum number of sequence letters per line in pairwise alignment.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Side effect
}
\description{
Report the final haplotype sequences.
}
\details{
\code{report} will create a \code{report} directory containing the consensus
sequences \code{map\{1,2,3\}.\{A,B\}.\{readtype\}.\{mapper\}.unchecked.fa},
an html alignment file \code{map\{1,2,3\}.aln.\{readtype\}.\{mapper\}.unchecked.html},
a pairwise alignment file \code{map\{1,2,3\}.\{A,B\}.\{readtype\}.\{mapper\}.unchecked.pair},
and a tab-separated file \code{problems.\{readtype\}.\{mapper\}.tsv} reporting
positions with potentially problematic results.

To finalise the consensus sequences open the pairwise alignment file and save
it replacing \strong{unchecked} with \strong{checked} in the file name.
In the \strong{checked} file manual edits can be performed. After the pairwise
alignment is verified call \code{\link{report_checked_consensus}}.
}
\examples{
\dontrun{
x <- DR2Smap(
  sample = "ID12300527",
  locus = "DPB1",
  datadir = "/path/to/data",
  outdir = "/path/to/output",
  reference = "04:01:01:01",
  consensus = "mapping"
  ) \%>\%
  clear() \%>\%
  mapInit() \%>\%
  partitionLongReads() \%>\%
  splitReadsByHaplotype() \%>\%
  extractFastq() \%>\%
  mapIter() \%>\%
  partitionShortReads() \%>\%
  mapFinal() \%>\%
  polish() \%>\%
  report(block_width = 60)
}
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extractFastq}}, \code{\link{mapFinal}},
  \code{\link{mapInit}}, \code{\link{mapIter}},
  \code{\link{partitionLongReads}}, \code{\link{polish}},
  \code{\link{report_checked_consensus}},
  \code{\link{splitReadsByHaplotype}}
}
