% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_checked_consensus}
\alias{report_checked_consensus}
\title{Report haplotype consensus files as verified}
\usage{
report_checked_consensus(x, which = "map3")
}
\arguments{
\item{x}{A \code{\link[=DR2S_]{DR2S}} object.}

\item{which}{Which mapping do we want to update.}
}
\value{
Side effect
}
\description{
Report haplotype consensus files as verified
}
\details{
\code{report_checked_consensus} will create a \code{checked} directory
containing the consensus sequences \code{\{A,B\}.\{readtype\}.\{mapper\}.checked.fa}
and a html alignment file \code{aln.\{readtype\}.\{mapper\}.checked.html}.
}
\seealso{
Other DR2S mapper functions: \code{\link{DR2Smap}},
  \code{\link{cache}}, \code{\link{clear}},
  \code{\link{extract_fastq}}, \code{\link{map0}},
  \code{\link{map1}}, \code{\link{map2}},
  \code{\link{map3}}, \code{\link{partition_haplotypes}},
  \code{\link{polish}}, \code{\link{report}},
  \code{\link{split_reads_by_haplotype}}
}
