% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-matrix.R
\name{SNPmatrix}
\alias{SNPmatrix}
\title{Construct a SNP matrix.}
\usage{
SNPmatrix(bamfile, max_depth, polymorphic_positions, cleanup = TRUE)
}
\arguments{
\item{bamfile}{A BAM file}

\item{max_depth}{How many reads do we consider.}

\item{polymorphic_positions}{SNP positions.}

\item{cleanup}{Remove temporary files.}
}
\value{
A \code{[read x position]} matrix.
}
\description{
Construct a SNP matrix.
}
\examples{
##
}
