% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-matrix.R
\name{SNPmatrix}
\alias{SNPmatrix}
\title{Construct a SNP matrix.}
\usage{
SNPmatrix(bamfile, refseq, polymorphic_positions)
}
\arguments{
\item{bamfile}{A BAM file}

\item{refseq}{Reference sequence as DNAString}

\item{polymorphic_positions}{SNP positions.}
}
\value{
A \code{[read x position]} matrix.
}
\description{
Construct a SNP matrix.
}
\examples{
##
}
