% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{writeDR2SConf}
\alias{writeDR2SConf}
\title{Write the DR2S config as yaml.}
\usage{
writeDR2SConf(x, outFile = NULL)
}
\arguments{
\item{x}{A \code{\link{DR2S}} object.}

\item{outFile}{The destination file.}
}
\description{
Write the DR2S config as yaml.
}
